+++ plugins/modules/docker_container.py
@@ -670,8 +670,6 @@ options:
   pid_mode:
     description:
       - Set the PID namespace mode for the container.
-      - Note that Docker SDK for Python < 2.0 only supports C(host). Newer versions of the
-        Docker SDK for Python (docker) allow all values supported by the Docker daemon.
     type: str
   pids_limit:
     description:
@@ -898,7 +896,6 @@ author:
   - "Felix Fontein (@felixfontein)"
 
 requirements:
-  - "L(Docker SDK for Python,https://docker-py.readthedocs.io/en/stable/) >= 1.8.0"
   - "Docker API >= 1.25"
 '''
 
@@ -1209,14 +1206,19 @@ import traceback
 from time import sleep
 
 from ansible.module_utils.common.text.formatters import human_to_bytes
-from ansible.module_utils.six import string_types
 from ansible.module_utils.common.text.converters import to_native, to_text
+from ansible.module_utils.six import string_types
 
 from ansible_collections.community.docker.plugins.module_utils.version import LooseVersion
 
-from ansible_collections.community.docker.plugins.module_utils.common import (
+from ansible_collections.community.docker.plugins.module_utils.common_api import (
     AnsibleDockerClient,
     RequestException,
+)
+from ansible_collections.community.docker.plugins.module_utils.module_container import (
+    DockerAPIEngineDriver,
+    OPTIONS,
+    Option,
 )
 from ansible_collections.community.docker.plugins.module_utils.util import (
     DifferenceTracker,
@@ -1230,951 +1232,22 @@ from ansible_collections.community.docker.plugins.module_utils.util import (
     DOCKER_COMMON_ARGS,
 )
 
-try:
-    from docker import utils
-    from ansible_collections.community.docker.plugins.module_utils.common import docker_version
-    if LooseVersion(docker_version) >= LooseVersion('1.10.0'):
-        from docker.types import Ulimit, LogConfig
-        from docker import types as docker_types
-    else:
-        from docker.utils.types import Ulimit, LogConfig
-    from docker.errors import DockerException, APIError, NotFound
-except Exception:
-    # missing Docker SDK for Python handled in ansible.module_utils.docker.common
-    pass
+from ansible_collections.community.docker.plugins.module_utils._api.errors import APIError, DockerException, NotFound
 
-
-REQUIRES_CONVERSION_TO_BYTES = [
-    'kernel_memory',
-    'memory',
-    'memory_reservation',
-    'memory_swap',
-    'shm_size'
-]
-
-
-def is_volume_permissions(mode):
-    for part in mode.split(','):
-        if part not in ('rw', 'ro', 'z', 'Z', 'consistent', 'delegated', 'cached', 'rprivate', 'private', 'rshared', 'shared', 'rslave', 'slave', 'nocopy'):
-            return False
-    return True
-
-
-def parse_port_range(range_or_port, client):
-    '''
-    Parses a string containing either a single port or a range of ports.
-
-    Returns a list of integers for each port in the list.
-    '''
-    if '-' in range_or_port:
-        try:
-            start, end = [int(port) for port in range_or_port.split('-')]
-        except Exception:
-            client.fail('Invalid port range: "{0}"'.format(range_or_port))
-        if end < start:
-            client.fail('Invalid port range: "{0}"'.format(range_or_port))
-        return list(range(start, end + 1))
-    else:
-        try:
-            return [int(range_or_port)]
-        except Exception:
-            client.fail('Invalid port: "{0}"'.format(range_or_port))
-
-
-def split_colon_ipv6(text, client):
-    '''
-    Split string by ':', while keeping IPv6 addresses in square brackets in one component.
-    '''
-    if '[' not in text:
-        return text.split(':')
-    start = 0
-    result = []
-    while start < len(text):
-        i = text.find('[', start)
-        if i < 0:
-            result.extend(text[start:].split(':'))
-            break
-        j = text.find(']', i)
-        if j < 0:
-            client.fail('Cannot find closing "]" in input "{0}" for opening "[" at index {1}!'.format(text, i + 1))
-        result.extend(text[start:i].split(':'))
-        k = text.find(':', j)
-        if k < 0:
-            result[-1] += text[i:]
-            start = len(text)
-        else:
-            result[-1] += text[i:k]
-            if k == len(text):
-                result.append('')
-                break
-            start = k + 1
-    return result
-
-
-class TaskParameters(DockerBaseClass):
-    '''
-    Access and parse module parameters
-    '''
-
-    def __init__(self, client):
-        super(TaskParameters, self).__init__()
-        self.client = client
-
-        self.auto_remove = None
-        self.blkio_weight = None
-        self.capabilities = None
-        self.cap_drop = None
-        self.cleanup = None
-        self.command = None
-        self.cpu_period = None
-        self.cpu_quota = None
-        self.cpus = None
-        self.cpuset_cpus = None
-        self.cpuset_mems = None
-        self.cpu_shares = None
-        self.debug = None
-        self.default_host_ip = None
-        self.detach = None
-        self.devices = None
-        self.device_read_bps = None
-        self.device_write_bps = None
-        self.device_read_iops = None
-        self.device_write_iops = None
-        self.device_requests = None
-        self.dns_servers = None
-        self.dns_opts = None
-        self.dns_search_domains = None
-        self.domainname = None
-        self.env = None
-        self.env_file = None
-        self.entrypoint = None
-        self.etc_hosts = None
-        self.exposed_ports = None
-        self.force_kill = None
-        self.groups = None
-        self.healthcheck = None
-        self.hostname = None
-        self.ignore_image = None
-        self.image = None
-        self.init = None
-        self.interactive = None
-        self.ipc_mode = None
-        self.keep_volumes = None
-        self.kernel_memory = None
-        self.kill_signal = None
-        self.labels = None
-        self.links = None
-        self.log_driver = None
-        self.output_logs = None
-        self.log_options = None
-        self.mac_address = None
-        self.memory = None
-        self.memory_reservation = None
-        self.memory_swap = None
-        self.memory_swappiness = None
-        self.mounts = None
-        self.name = None
-        self.network_mode = None
-        self.userns_mode = None
-        self.networks = None
-        self.networks_cli_compatible = None
-        self.oom_killer = None
-        self.oom_score_adj = None
-        self.paused = None
-        self.pid_mode = None
-        self.pids_limit = None
-        self.privileged = None
-        self.purge_networks = None
-        self.pull = None
-        self.read_only = None
-        self.recreate = None
-        self.removal_wait_timeout = None
-        self.restart = None
-        self.restart_retries = None
-        self.restart_policy = None
-        self.runtime = None
-        self.shm_size = None
-        self.security_opts = None
-        self.state = None
-        self.stop_signal = None
-        self.stop_timeout = None
-        self.storage_opts = None
-        self.tmpfs = None
-        self.tty = None
-        self.user = None
-        self.uts = None
-        self.volumes = None
-        self.volume_binds = dict()
-        self.volumes_from = None
-        self.volume_driver = None
-        self.working_dir = None
-
-        for key, value in client.module.params.items():
-            setattr(self, key, value)
-        self.comparisons = client.comparisons
-
-        # If state is 'absent', parameters do not have to be parsed or interpreted.
-        # Only the container's name is needed.
-        if self.state == 'absent':
-            return
-
-        if self.default_host_ip:
-            valid_ip = False
-            if re.match(r'^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$', self.default_host_ip):
-                valid_ip = True
-            if re.match(r'^\[[0-9a-fA-F:]+\]$', self.default_host_ip):
-                valid_ip = True
-            if re.match(r'^[0-9a-fA-F:]+$', self.default_host_ip):
-                self.default_host_ip = '[{0}]'.format(self.default_host_ip)
-                valid_ip = True
-            if not valid_ip:
-                self.fail('The value of default_host_ip must be an empty string, an IPv4 address, '
-                          'or an IPv6 address. Got "{0}" instead.'.format(self.default_host_ip))
-
-        if self.cpus is not None:
-            self.cpus = int(round(self.cpus * 1E9))
-
-        if self.groups:
-            # In case integers are passed as groups, we need to convert them to
-            # strings as docker internally treats them as strings.
-            self.groups = [to_text(g, errors='surrogate_or_strict') for g in self.groups]
-
-        for param_name in REQUIRES_CONVERSION_TO_BYTES:
-            if client.module.params.get(param_name):
-                if param_name == 'memory_swap' and client.module.params.get(param_name) in ['unlimited', '-1']:
-                    setattr(self, param_name, -1)
-                else:
-                    try:
-                        setattr(self, param_name, human_to_bytes(client.module.params.get(param_name)))
-                    except ValueError as exc:
-                        self.fail("Failed to convert %s to bytes: %s" % (param_name, to_native(exc)))
-
-        self.published_ports = self._parse_publish_ports()
-
-        self.ports = self._parse_exposed_ports(self.published_ports)
-        self.log("expose ports:")
-        self.log(self.ports, pretty_print=True)
-
-        self.links = self._parse_links(self.links)
-
-        if self.volumes:
-            self.volumes = self._expand_host_paths()
-
-        self.tmpfs = self._parse_tmpfs()
-        self.env = self._get_environment()
-        self.ulimits = self._parse_ulimits()
-        self.sysctls = self._parse_sysctls()
-        self.log_config = self._parse_log_config()
-        try:
-            self.healthcheck, self.disable_healthcheck = parse_healthcheck(self.healthcheck)
-        except ValueError as e:
-            self.fail(to_native(e))
-
-        self.exp_links = None
-        self.volume_binds = self._get_volume_binds(self.volumes)
-        self.pid_mode = self._replace_container_names(self.pid_mode)
-        self.ipc_mode = self._replace_container_names(self.ipc_mode)
-        self.network_mode = self._replace_container_names(self.network_mode)
-
-        self.log("volumes:")
-        self.log(self.volumes, pretty_print=True)
-        self.log("volume binds:")
-        self.log(self.volume_binds, pretty_print=True)
-
-        if self.networks:
-            for network in self.networks:
-                network['id'] = self._get_network_id(network['name'])
-                if not network['id']:
-                    self.fail("Parameter error: network named %s could not be found. Does it exist?" % network['name'])
-                if network.get('links'):
-                    network['links'] = self._parse_links(network['links'])
-
-        if self.mac_address:
-            # Ensure the MAC address uses colons instead of hyphens for later comparison
-            self.mac_address = self.mac_address.replace('-', ':')
-
-        if client.module.params['command_handling'] == 'correct':
-            if self.entrypoint is not None:
-                self.entrypoint = [to_text(x, errors='surrogate_or_strict') for x in self.entrypoint]
-
-            if self.command is not None:
-                if not isinstance(self.command, list):
-                    # convert from str to list
-                    self.command = shlex.split(to_text(self.command, errors='surrogate_or_strict'))
-                self.command = [to_text(x, errors='surrogate_or_strict') for x in self.command]
-        else:
-            if self.entrypoint:
-                # convert from list to str.
-                self.entrypoint = shlex.split(' '.join([to_text(x, errors='surrogate_or_strict') for x in self.entrypoint]))
-                self.entrypoint = [to_text(x, errors='surrogate_or_strict') for x in self.entrypoint]
-
-            if self.command:
-                # convert from list to str
-                if isinstance(self.command, list):
-                    self.command = ' '.join([to_text(x, errors='surrogate_or_strict') for x in self.command])
-                else:
-                    self.command = to_text(self.command, errors='surrogate_or_strict')
-                self.command = [to_text(x, errors='surrogate_or_strict') for x in shlex.split(self.command)]
-
-        self.mounts_opt, self.expected_mounts = self._process_mounts()
-
-        self._check_mount_target_collisions()
-
-        for param_name in ["device_read_bps", "device_write_bps"]:
-            if client.module.params.get(param_name):
-                self._process_rate_bps(option=param_name)
-
-        for param_name in ["device_read_iops", "device_write_iops"]:
-            if client.module.params.get(param_name):
-                self._process_rate_iops(option=param_name)
-
-        if self.device_requests:
-            for dr_index, dr in enumerate(self.device_requests):
-                # Make sure that capabilities are lists of lists of strings
-                if dr['capabilities']:
-                    for or_index, or_list in enumerate(dr['capabilities']):
-                        for and_index, and_term in enumerate(or_list):
-                            if not isinstance(and_term, string_types):
-                                self.fail(
-                                    "device_requests[{0}].capabilities[{1}][{2}] is not a string".format(
-                                        dr_index, or_index, and_index))
-                            or_list[and_index] = to_native(and_term)
-                # Make sure that options is a dictionary mapping strings to strings
-                if dr['options']:
-                    dr['options'] = clean_dict_booleans_for_docker_api(dr['options'])
-
-    def fail(self, msg):
-        self.client.fail(msg)
-
-    @property
-    def update_parameters(self):
-        '''
-        Returns parameters used to update a container
-        '''
-
-        update_parameters = dict(
-            blkio_weight='blkio_weight',
-            cpu_period='cpu_period',
-            cpu_quota='cpu_quota',
-            cpu_shares='cpu_shares',
-            cpuset_cpus='cpuset_cpus',
-            cpuset_mems='cpuset_mems',
-            mem_limit='memory',
-            mem_reservation='memory_reservation',
-            memswap_limit='memory_swap',
-            kernel_memory='kernel_memory',
-            restart_policy='restart_policy',
-        )
-
-        result = dict()
-        for key, value in update_parameters.items():
-            if getattr(self, value, None) is not None:
-                if key == 'restart_policy' and self.client.option_minimal_versions[value]['supported']:
-                    restart_policy = dict(Name=self.restart_policy,
-                                          MaximumRetryCount=self.restart_retries)
-                    result[key] = restart_policy
-                elif self.client.option_minimal_versions[value]['supported']:
-                    result[key] = getattr(self, value)
-        return result
-
-    @property
-    def create_parameters(self):
-        '''
-        Returns parameters used to create a container
-        '''
-        create_params = dict(
-            command='command',
-            domainname='domainname',
-            hostname='hostname',
-            user='user',
-            detach='detach',
-            stdin_open='interactive',
-            tty='tty',
-            ports='ports',
-            environment='env',
-            name='name',
-            entrypoint='entrypoint',
-            mac_address='mac_address',
-            labels='labels',
-            stop_signal='stop_signal',
-            working_dir='working_dir',
-            stop_timeout='stop_timeout',
-            healthcheck='healthcheck',
-        )
-
-        if self.client.docker_py_version < LooseVersion('3.0'):
-            # cpu_shares and volume_driver moved to create_host_config in > 3
-            create_params['cpu_shares'] = 'cpu_shares'
-            create_params['volume_driver'] = 'volume_driver'
-
-        result = dict(
-            host_config=self._host_config(),
-            volumes=self._get_mounts(),
-        )
-
-        for key, value in create_params.items():
-            if getattr(self, value, None) is not None:
-                if self.client.option_minimal_versions[value]['supported']:
-                    result[key] = getattr(self, value)
-
-        if self.disable_healthcheck:
-            # Make sure image's health check is overridden
-            result['healthcheck'] = {'test': ['NONE']}
-
-        if self.networks_cli_compatible and self.networks:
-            network = self.networks[0]
-            params = dict()
-            for para in ('ipv4_address', 'ipv6_address', 'links', 'aliases'):
-                if network.get(para):
-                    params[para] = network[para]
-            network_config = dict()
-            network_config[network['name']] = self.client.create_endpoint_config(**params)
-            result['networking_config'] = self.client.create_networking_config(network_config)
-        return result
-
-    def _expand_host_paths(self):
-        new_vols = []
-        for vol in self.volumes:
-            if ':' in vol:
-                parts = vol.split(':')
-                if len(parts) == 3:
-                    host, container, mode = parts
-                    if not is_volume_permissions(mode):
-                        self.fail('Found invalid volumes mode: {0}'.format(mode))
-                    if re.match(r'[.~]', host):
-                        host = os.path.abspath(os.path.expanduser(host))
-                    new_vols.append("%s:%s:%s" % (host, container, mode))
-                    continue
-                elif len(parts) == 2:
-                    if not is_volume_permissions(parts[1]) and re.match(r'[.~]', parts[0]):
-                        host = os.path.abspath(os.path.expanduser(parts[0]))
-                        new_vols.append("%s:%s:rw" % (host, parts[1]))
-                        continue
-            new_vols.append(vol)
-        return new_vols
-
-    def _get_mounts(self):
-        '''
-        Return a list of container mounts.
-        :return:
-        '''
-        result = []
-        if self.volumes:
-            for vol in self.volumes:
-                # Only pass anonymous volumes to create container
-                if ':' in vol:
-                    parts = vol.split(':')
-                    if len(parts) == 3:
-                        continue
-                    if len(parts) == 2:
-                        if not is_volume_permissions(parts[1]):
-                            continue
-                result.append(vol)
-        self.log("mounts:")
-        self.log(result, pretty_print=True)
-        return result
-
-    def _host_config(self):
-        '''
-        Returns parameters used to create a HostConfig object
-        '''
-
-        host_config_params = dict(
-            port_bindings='published_ports',
-            publish_all_ports='publish_all_ports',
-            links='links',
-            privileged='privileged',
-            cgroup_parent='cgroup_parent',
-            dns='dns_servers',
-            dns_opt='dns_opts',
-            dns_search='dns_search_domains',
-            binds='volume_binds',
-            volumes_from='volumes_from',
-            network_mode='network_mode',
-            userns_mode='userns_mode',
-            cap_add='capabilities',
-            cap_drop='cap_drop',
-            extra_hosts='etc_hosts',
-            read_only='read_only',
-            ipc_mode='ipc_mode',
-            security_opt='security_opts',
-            ulimits='ulimits',
-            sysctls='sysctls',
-            log_config='log_config',
-            mem_limit='memory',
-            memswap_limit='memory_swap',
-            mem_swappiness='memory_swappiness',
-            oom_score_adj='oom_score_adj',
-            oom_kill_disable='oom_killer',
-            shm_size='shm_size',
-            group_add='groups',
-            devices='devices',
-            pid_mode='pid_mode',
-            tmpfs='tmpfs',
-            init='init',
-            uts_mode='uts',
-            runtime='runtime',
-            auto_remove='auto_remove',
-            device_read_bps='device_read_bps',
-            device_write_bps='device_write_bps',
-            device_read_iops='device_read_iops',
-            device_write_iops='device_write_iops',
-            pids_limit='pids_limit',
-            mounts='mounts',
-            nano_cpus='cpus',
-            storage_opt='storage_opts',
-        )
-
-        if self.client.docker_py_version >= LooseVersion('1.9'):
-            # blkio_weight can always be updated, but can only be set on creation
-            # when Docker SDK for Python and Docker API are new enough
-            host_config_params['blkio_weight'] = 'blkio_weight'
-
-        if self.client.docker_py_version >= LooseVersion('3.0'):
-            # cpu_shares and volume_driver moved to create_host_config in > 3
-            host_config_params['cpu_shares'] = 'cpu_shares'
-            host_config_params['volume_driver'] = 'volume_driver'
-
-        params = dict()
-        for key, value in host_config_params.items():
-            if getattr(self, value, None) is not None:
-                if self.client.option_minimal_versions[value]['supported']:
-                    params[key] = getattr(self, value)
-
-        if self.restart_policy:
-            params['restart_policy'] = dict(Name=self.restart_policy,
-                                            MaximumRetryCount=self.restart_retries)
-
-        if 'mounts' in params:
-            params['mounts'] = self.mounts_opt
-
-        if self.device_requests is not None:
-            params['device_requests'] = [dict((k, v) for k, v in dr.items() if v is not None) for dr in self.device_requests]
-
-        return self.client.create_host_config(**params)
-
-    def get_default_host_ip(self):
-        if self.default_host_ip is not None:
-            return self.default_host_ip
-        ip = '0.0.0.0'
-        if not self.networks:
-            return ip
-        for net in self.networks:
-            if net.get('name'):
-                try:
-                    network = self.client.inspect_network(net['name'])
-                    if network.get('Driver') == 'bridge' and \
-                       network.get('Options', {}).get('com.docker.network.bridge.host_binding_ipv4'):
-                        ip = network['Options']['com.docker.network.bridge.host_binding_ipv4']
-                        break
-                except NotFound as nfe:
-                    self.client.fail(
-                        "Cannot inspect the network '{0}' to determine the default IP: {1}".format(net['name'], to_native(nfe)),
-                        exception=traceback.format_exc()
-                    )
-        return ip
-
-    def _parse_publish_ports(self):
-        '''
-        Parse ports from docker CLI syntax
-        '''
-        if self.published_ports is None:
-            return None
-
-        if 'all' in self.published_ports:
-            self.client.module.fail_json(
-                msg='Specifying "all" in published_ports is no longer allowed. Set publish_all_ports to "true" instead '
-                    'to randomly assign port mappings for those not specified by published_ports.')
-
-        default_ip = self.get_default_host_ip()
-
-        binds = {}
-        for port in self.published_ports:
-            parts = split_colon_ipv6(to_text(port, errors='surrogate_or_strict'), self.client)
-            container_port = parts[-1]
-            protocol = ''
-            if '/' in container_port:
-                container_port, protocol = parts[-1].split('/')
-            container_ports = parse_port_range(container_port, self.client)
-
-            p_len = len(parts)
-            if p_len == 1:
-                port_binds = len(container_ports) * [(default_ip,)]
-            elif p_len == 2:
-                if len(container_ports) == 1:
-                    port_binds = [(default_ip, parts[0])]
-                else:
-                    port_binds = [(default_ip, port) for port in parse_port_range(parts[0], self.client)]
-            elif p_len == 3:
-                # We only allow IPv4 and IPv6 addresses for the bind address
-                ipaddr = parts[0]
-                if not re.match(r'^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$', parts[0]) and not re.match(r'^\[[0-9a-fA-F:]+(?:|%[^\]/]+)\]$', ipaddr):
-                    self.fail(('Bind addresses for published ports must be IPv4 or IPv6 addresses, not hostnames. '
-                               'Use the dig lookup to resolve hostnames. (Found hostname: {0})').format(ipaddr))
-                if re.match(r'^\[[0-9a-fA-F:]+\]$', ipaddr):
-                    ipaddr = ipaddr[1:-1]
-                if parts[1]:
-                    if len(container_ports) == 1:
-                        port_binds = [(ipaddr, parts[1])]
-                    else:
-                        port_binds = [(ipaddr, port) for port in parse_port_range(parts[1], self.client)]
-                else:
-                    port_binds = len(container_ports) * [(ipaddr,)]
-            else:
-                self.fail(('Invalid port description "%s" - expected 1 to 3 colon-separated parts, but got %d. '
-                           'Maybe you forgot to use square brackets ([...]) around an IPv6 address?') % (port, p_len))
-
-            for bind, container_port in zip(port_binds, container_ports):
-                idx = '{0}/{1}'.format(container_port, protocol) if protocol else container_port
-                if idx in binds:
-                    old_bind = binds[idx]
-                    if isinstance(old_bind, list):
-                        old_bind.append(bind)
-                    else:
-                        binds[idx] = [old_bind, bind]
-                else:
-                    binds[idx] = bind
-        return binds
-
-    def _get_volume_binds(self, volumes):
-        '''
-        Extract host bindings, if any, from list of volume mapping strings.
-
-        :return: dictionary of bind mappings
-        '''
-        result = dict()
-        if volumes:
-            for vol in volumes:
-                host = None
-                if ':' in vol:
-                    parts = vol.split(':')
-                    if len(parts) == 3:
-                        host, container, mode = parts
-                        if not is_volume_permissions(mode):
-                            self.fail('Found invalid volumes mode: {0}'.format(mode))
-                    elif len(parts) == 2:
-                        if not is_volume_permissions(parts[1]):
-                            host, container, mode = (parts + ['rw'])
-                if host is not None:
-                    result[host] = dict(
-                        bind=container,
-                        mode=mode
-                    )
-        return result
-
-    def _parse_exposed_ports(self, published_ports):
-        '''
-        Parse exposed ports from docker CLI-style ports syntax.
-        '''
-        exposed = []
-        if self.exposed_ports:
-            for port in self.exposed_ports:
-                port = to_text(port, errors='surrogate_or_strict').strip()
-                protocol = 'tcp'
-                match = re.search(r'(/.+$)', port)
-                if match:
-                    protocol = match.group(1).replace('/', '')
-                    port = re.sub(r'/.+$', '', port)
-                exposed.append((port, protocol))
-        if published_ports:
-            # Any published port should also be exposed
-            for publish_port in published_ports:
-                match = False
-                if isinstance(publish_port, string_types) and '/' in publish_port:
-                    port, protocol = publish_port.split('/')
-                    port = int(port)
-                else:
-                    protocol = 'tcp'
-                    port = int(publish_port)
-                for exposed_port in exposed:
-                    if exposed_port[1] != protocol:
-                        continue
-                    if isinstance(exposed_port[0], string_types) and '-' in exposed_port[0]:
-                        start_port, end_port = exposed_port[0].split('-')
-                        if int(start_port) <= port <= int(end_port):
-                            match = True
-                    elif exposed_port[0] == port:
-                        match = True
-                if not match:
-                    exposed.append((port, protocol))
-        return exposed
-
-    @staticmethod
-    def _parse_links(links):
-        '''
-        Turn links into a dictionary
-        '''
-        if links is None:
-            return None
-
-        result = []
-        for link in links:
-            parsed_link = link.split(':', 1)
-            if len(parsed_link) == 2:
-                result.append((parsed_link[0], parsed_link[1]))
-            else:
-                result.append((parsed_link[0], parsed_link[0]))
-        return result
-
-    def _parse_ulimits(self):
-        '''
-        Turn ulimits into an array of Ulimit objects
-        '''
-        if self.ulimits is None:
-            return None
-
-        results = []
-        for limit in self.ulimits:
-            limits = dict()
-            pieces = limit.split(':')
-            if len(pieces) >= 2:
-                limits['name'] = pieces[0]
-                limits['soft'] = int(pieces[1])
-                limits['hard'] = int(pieces[1])
-            if len(pieces) == 3:
-                limits['hard'] = int(pieces[2])
-            try:
-                results.append(Ulimit(**limits))
-            except ValueError as exc:
-                self.fail("Error parsing ulimits value %s - %s" % (limit, to_native(exc)))
-        return results
-
-    def _parse_sysctls(self):
-        '''
-        Turn sysctls into an hash of Sysctl objects
-        '''
-        return self.sysctls
-
-    def _parse_log_config(self):
-        '''
-        Create a LogConfig object
-        '''
-        if self.log_driver is None:
-            return None
-
-        options = dict(
-            Type=self.log_driver,
-            Config=dict()
-        )
-
-        if self.log_options is not None:
-            options['Config'] = dict()
-            for k, v in self.log_options.items():
-                if not isinstance(v, string_types):
-                    self.client.module.warn(
-                        "Non-string value found for log_options option '%s'. The value is automatically converted to '%s'. "
-                        "If this is not correct, or you want to avoid such warnings, please quote the value." % (
-                            k, to_text(v, errors='surrogate_or_strict'))
-                    )
-                v = to_text(v, errors='surrogate_or_strict')
-                self.log_options[k] = v
-                options['Config'][k] = v
-
-        try:
-            return LogConfig(**options)
-        except ValueError as exc:
-            self.fail('Error parsing logging options - %s' % (to_native(exc), ))
-
-    def _parse_tmpfs(self):
-        '''
-        Turn tmpfs into a hash of Tmpfs objects
-        '''
-        result = dict()
-        if self.tmpfs is None:
-            return result
-
-        for tmpfs_spec in self.tmpfs:
-            split_spec = tmpfs_spec.split(":", 1)
-            if len(split_spec) > 1:
-                result[split_spec[0]] = split_spec[1]
-            else:
-                result[split_spec[0]] = ""
-        return result
-
-    def _get_environment(self):
-        """
-        If environment file is combined with explicit environment variables, the explicit environment variables
-        take precedence.
-        """
-        final_env = {}
-        if self.env_file:
-            parsed_env_file = utils.parse_env_file(self.env_file)
-            for name, value in parsed_env_file.items():
-                final_env[name] = to_text(value, errors='surrogate_or_strict')
-        if self.env:
-            for name, value in self.env.items():
-                if not isinstance(value, string_types):
-                    self.fail("Non-string value found for env option. Ambiguous env options must be "
-                              "wrapped in quotes to avoid them being interpreted. Key: %s" % (name, ))
-                final_env[name] = to_text(value, errors='surrogate_or_strict')
-        return final_env
-
-    def _get_network_id(self, network_name):
-        network_id = None
-        try:
-            for network in self.client.networks(names=[network_name]):
-                if network['Name'] == network_name:
-                    network_id = network['Id']
-                    break
-        except Exception as exc:
-            self.fail("Error getting network id for %s - %s" % (network_name, to_native(exc)))
-        return network_id
-
-    def _process_mounts(self):
-        if self.mounts is None:
-            return None, None
-        mounts_list = []
-        mounts_expected = []
-        for mount in self.mounts:
-            target = mount['target']
-            datatype = mount['type']
-            mount_dict = dict(mount)
-            # Sanity checks (so we don't wait for Docker SDK for Python to barf on input)
-            if mount_dict.get('source') is None and datatype not in ('tmpfs', 'volume'):
-                self.client.fail('source must be specified for mount "{0}" of type "{1}"'.format(target, datatype))
-            mount_option_types = dict(
-                volume_driver='volume',
-                volume_options='volume',
-                propagation='bind',
-                no_copy='volume',
-                labels='volume',
-                tmpfs_size='tmpfs',
-                tmpfs_mode='tmpfs',
-            )
-            for option, req_datatype in mount_option_types.items():
-                if mount_dict.get(option) is not None and datatype != req_datatype:
-                    self.client.fail('{0} cannot be specified for mount "{1}" of type "{2}" (needs type "{3}")'.format(option, target, datatype, req_datatype))
-            # Handle volume_driver and volume_options
-            volume_driver = mount_dict.pop('volume_driver')
-            volume_options = mount_dict.pop('volume_options')
-            if volume_driver:
-                if volume_options:
-                    volume_options = clean_dict_booleans_for_docker_api(volume_options)
-                mount_dict['driver_config'] = docker_types.DriverConfig(name=volume_driver, options=volume_options)
-            if mount_dict['labels']:
-                mount_dict['labels'] = clean_dict_booleans_for_docker_api(mount_dict['labels'])
-            if mount_dict.get('tmpfs_size') is not None:
-                try:
-                    mount_dict['tmpfs_size'] = human_to_bytes(mount_dict['tmpfs_size'])
-                except ValueError as exc:
-                    self.fail('Failed to convert tmpfs_size of mount "{0}" to bytes: {1}'.format(target, to_native(exc)))
-            if mount_dict.get('tmpfs_mode') is not None:
-                try:
-                    mount_dict['tmpfs_mode'] = int(mount_dict['tmpfs_mode'], 8)
-                except Exception as dummy:
-                    self.client.fail('tmp_fs mode of mount "{0}" is not an octal string!'.format(target))
-            # Fill expected mount dict
-            mount_expected = dict(mount)
-            mount_expected['tmpfs_size'] = mount_dict['tmpfs_size']
-            mount_expected['tmpfs_mode'] = mount_dict['tmpfs_mode']
-            # Add result to lists
-            mounts_list.append(docker_types.Mount(**mount_dict))
-            mounts_expected.append(omit_none_from_dict(mount_expected))
-        return mounts_list, mounts_expected
-
-    def _process_rate_bps(self, option):
-        """
-        Format device_read_bps and device_write_bps option
-        """
-        devices_list = []
-        for v in getattr(self, option):
-            device_dict = dict((x.title(), y) for x, y in v.items())
-            device_dict['Rate'] = human_to_bytes(device_dict['Rate'])
-            devices_list.append(device_dict)
-
-        setattr(self, option, devices_list)
-
-    def _process_rate_iops(self, option):
-        """
-        Format device_read_iops and device_write_iops option
-        """
-        devices_list = []
-        for v in getattr(self, option):
-            device_dict = dict((x.title(), y) for x, y in v.items())
-            devices_list.append(device_dict)
-
-        setattr(self, option, devices_list)
-
-    def _replace_container_names(self, mode):
-        """
-        Parse IPC and PID modes. If they contain a container name, replace
-        with the container's ID.
-        """
-        if mode is None or not mode.startswith('container:'):
-            return mode
-        container_name = mode[len('container:'):]
-        # Try to inspect container to see whether this is an ID or a
-        # name (and in the latter case, retrieve it's ID)
-        container = self.client.get_container(container_name)
-        if container is None:
-            # If we can't find the container, issue a warning and continue with
-            # what the user specified.
-            self.client.module.warn('Cannot find a container with name or ID "{0}"'.format(container_name))
-            return mode
-        return 'container:{0}'.format(container['Id'])
-
-    def _check_mount_target_collisions(self):
-        last = dict()
-
-        def f(t, name):
-            if t in last:
-                if name == last[t]:
-                    self.client.fail('The mount point "{0}" appears twice in the {1} option'.format(t, name))
-                else:
-                    self.client.fail('The mount point "{0}" appears both in the {1} and {2} option'.format(t, name, last[t]))
-            last[t] = name
-
-        if self.expected_mounts:
-            for t in [m['target'] for m in self.expected_mounts]:
-                f(t, 'mounts')
-        if self.volumes:
-            for v in self.volumes:
-                vs = v.split(':')
-                f(vs[0 if len(vs) == 1 else 1], 'volumes')
+from ansible_collections.community.docker.plugins.module_utils._api.utils.utils import parse_repository_tag, normalize_links
 
 
 class Container(DockerBaseClass):
-
-    def __init__(self, container, parameters):
+    def __init__(self, container):
         super(Container, self).__init__()
         self.raw = container
         self.Id = None
+        self.Image = None
         self.container = container
         if container:
             self.Id = container['Id']
             self.Image = container['Image']
         self.log(self.container, pretty_print=True)
-        self.parameters = parameters
-        self.parameters.expected_links = None
-        self.parameters.expected_ports = None
-        self.parameters.expected_exposed = None
-        self.parameters.expected_volumes = None
-        self.parameters.expected_ulimits = None
-        self.parameters.expected_sysctls = None
-        self.parameters.expected_etc_hosts = None
-        self.parameters.expected_env = None
-        self.parameters.expected_device_requests = None
-        self.parameters_map = dict()
-        self.parameters_map['expected_labels'] = 'labels'
-        self.parameters_map['expected_links'] = 'links'
-        self.parameters_map['expected_ports'] = 'expected_ports'
-        self.parameters_map['expected_exposed'] = 'exposed_ports'
-        self.parameters_map['expected_volumes'] = 'volumes'
-        self.parameters_map['expected_ulimits'] = 'ulimits'
-        self.parameters_map['expected_sysctls'] = 'sysctls'
-        self.parameters_map['expected_etc_hosts'] = 'etc_hosts'
-        self.parameters_map['expected_env'] = 'env'
-        self.parameters_map['expected_entrypoint'] = 'entrypoint'
-        self.parameters_map['expected_binds'] = 'volumes'
-        self.parameters_map['expected_labels'] = 'labels'
-        self.parameters_map['expected_cmd'] = 'command'
-        self.parameters_map['expected_devices'] = 'devices'
-        self.parameters_map['expected_healthcheck'] = 'healthcheck'
-        self.parameters_map['expected_mounts'] = 'mounts'
-        self.parameters_map['expected_device_requests'] = 'device_requests'
-
-    def fail(self, msg):
-        self.parameters.client.fail(msg)
 
     @property
     def exists(self):
@@ -2199,619 +1272,187 @@ class Container(DockerBaseClass):
             return self.container['State'].get('Paused', False)
         return False
 
-    def _compare(self, a, b, compare):
-        '''
-        Compare values a and b as described in compare.
-        '''
-        return compare_generic(a, b, compare['comparison'], compare['type'])
-
-    def _decode_mounts(self, mounts):
-        if not mounts:
-            return mounts
-        result = []
-        empty_dict = dict()
-        for mount in mounts:
-            res = dict()
-            res['type'] = mount.get('Type')
-            res['source'] = mount.get('Source')
-            res['target'] = mount.get('Target')
-            res['read_only'] = mount.get('ReadOnly', False)  # golang's omitempty for bool returns None for False
-            res['consistency'] = mount.get('Consistency')
-            res['propagation'] = mount.get('BindOptions', empty_dict).get('Propagation')
-            res['no_copy'] = mount.get('VolumeOptions', empty_dict).get('NoCopy', False)
-            res['labels'] = mount.get('VolumeOptions', empty_dict).get('Labels', empty_dict)
-            res['volume_driver'] = mount.get('VolumeOptions', empty_dict).get('DriverConfig', empty_dict).get('Name')
-            res['volume_options'] = mount.get('VolumeOptions', empty_dict).get('DriverConfig', empty_dict).get('Options', empty_dict)
-            res['tmpfs_size'] = mount.get('TmpfsOptions', empty_dict).get('SizeBytes')
-            res['tmpfs_mode'] = mount.get('TmpfsOptions', empty_dict).get('Mode')
-            result.append(res)
-        return result
-
-    def has_different_configuration(self, image):
-        '''
-        Diff parameters vs existing container config. Returns tuple: (True | False, List of differences)
-        '''
-        self.log('Starting has_different_configuration')
-        self.parameters.expected_entrypoint = self._get_expected_entrypoint()
-        self.parameters.expected_links = self._get_expected_links()
-        self.parameters.expected_ports = self._get_expected_ports()
-        self.parameters.expected_exposed = self._get_expected_exposed(image)
-        self.parameters.expected_volumes = self._get_expected_volumes(image)
-        self.parameters.expected_binds = self._get_expected_binds(image)
-        self.parameters.expected_labels = self._get_expected_labels(image)
-        self.parameters.expected_ulimits = self._get_expected_ulimits(self.parameters.ulimits)
-        self.parameters.expected_sysctls = self._get_expected_sysctls(self.parameters.sysctls)
-        self.parameters.expected_etc_hosts = self._convert_simple_dict_to_list('etc_hosts')
-        self.parameters.expected_env = self._get_expected_env(image)
-        self.parameters.expected_cmd = self._get_expected_cmd()
-        self.parameters.expected_devices = self._get_expected_devices()
-        self.parameters.expected_healthcheck = self._get_expected_healthcheck()
-        self.parameters.expected_device_requests = self._get_expected_device_requests()
-
-        if not self.container.get('HostConfig'):
-            self.fail("has_config_diff: Error parsing container properties. HostConfig missing.")
-        if not self.container.get('Config'):
-            self.fail("has_config_diff: Error parsing container properties. Config missing.")
-        if not self.container.get('NetworkSettings'):
-            self.fail("has_config_diff: Error parsing container properties. NetworkSettings missing.")
-
-        host_config = self.container['HostConfig']
-        log_config = host_config.get('LogConfig', dict())
-        config = self.container['Config']
-        network = self.container['NetworkSettings']
-
-        # The previous version of the docker module ignored the detach state by
-        # assuming if the container was running, it must have been detached.
-        detach = not (config.get('AttachStderr') and config.get('AttachStdout'))
-
-        # "ExposedPorts": null returns None type & causes AttributeError - PR #5517
-        if config.get('ExposedPorts') is not None:
-            expected_exposed = [self._normalize_port(p) for p in config.get('ExposedPorts', dict()).keys()]
-        else:
-            expected_exposed = []
-
-        # Map parameters to container inspect results
-        config_mapping = dict(
-            expected_cmd=config.get('Cmd'),
-            domainname=config.get('Domainname'),
-            hostname=config.get('Hostname'),
-            user=config.get('User'),
-            detach=detach,
-            init=host_config.get('Init'),
-            interactive=config.get('OpenStdin'),
-            capabilities=host_config.get('CapAdd'),
-            cap_drop=host_config.get('CapDrop'),
-            cgroup_parent=host_config.get('CgroupParent'),
-            expected_devices=host_config.get('Devices'),
-            dns_servers=host_config.get('Dns'),
-            dns_opts=host_config.get('DnsOptions'),
-            dns_search_domains=host_config.get('DnsSearch'),
-            expected_env=(config.get('Env') or []),
-            expected_entrypoint=config.get('Entrypoint'),
-            expected_etc_hosts=host_config['ExtraHosts'],
-            expected_exposed=expected_exposed,
-            groups=host_config.get('GroupAdd'),
-            ipc_mode=host_config.get("IpcMode"),
-            expected_labels=config.get('Labels'),
-            expected_links=host_config.get('Links'),
-            mac_address=config.get('MacAddress', network.get('MacAddress')),
-            memory_swappiness=host_config.get('MemorySwappiness'),
-            network_mode=host_config.get('NetworkMode'),
-            userns_mode=host_config.get('UsernsMode'),
-            oom_killer=host_config.get('OomKillDisable'),
-            oom_score_adj=host_config.get('OomScoreAdj'),
-            pid_mode=host_config.get('PidMode'),
-            privileged=host_config.get('Privileged'),
-            expected_ports=host_config.get('PortBindings'),
-            read_only=host_config.get('ReadonlyRootfs'),
-            runtime=host_config.get('Runtime'),
-            shm_size=host_config.get('ShmSize'),
-            security_opts=host_config.get("SecurityOpt"),
-            stop_signal=config.get("StopSignal"),
-            tmpfs=host_config.get('Tmpfs'),
-            tty=config.get('Tty'),
-            expected_ulimits=host_config.get('Ulimits'),
-            expected_sysctls=host_config.get('Sysctls'),
-            uts=host_config.get('UTSMode'),
-            expected_volumes=config.get('Volumes'),
-            expected_binds=host_config.get('Binds'),
-            volume_driver=host_config.get('VolumeDriver'),
-            volumes_from=host_config.get('VolumesFrom'),
-            working_dir=config.get('WorkingDir'),
-            publish_all_ports=host_config.get('PublishAllPorts'),
-            expected_healthcheck=config.get('Healthcheck'),
-            disable_healthcheck=(not config.get('Healthcheck') or config.get('Healthcheck').get('Test') == ['NONE']),
-            device_read_bps=host_config.get('BlkioDeviceReadBps'),
-            device_write_bps=host_config.get('BlkioDeviceWriteBps'),
-            device_read_iops=host_config.get('BlkioDeviceReadIOps'),
-            device_write_iops=host_config.get('BlkioDeviceWriteIOps'),
-            expected_device_requests=host_config.get('DeviceRequests'),
-            pids_limit=host_config.get('PidsLimit'),
-            storage_opts=host_config.get('StorageOpt'),
-            # According to https://github.com/moby/moby/, support for HostConfig.Mounts
-            # has been included at least since v17.03.0-ce, which has API version 1.26.
-            # The previous tag, v1.9.1, has API version 1.21 and does not have
-            # HostConfig.Mounts. I have no idea what about API 1.25...
-            expected_mounts=self._decode_mounts(host_config.get('Mounts')),
-            cpus=host_config.get('NanoCpus'),
-        )
-        # Options which don't make sense without their accompanying option
-        if self.parameters.log_driver:
-            config_mapping['log_driver'] = log_config.get('Type')
-            config_mapping['log_options'] = log_config.get('Config')
-
-        if self.parameters.client.option_minimal_versions['auto_remove']['supported']:
-            # auto_remove is only supported in Docker SDK for Python >= 2.0.0; unfortunately
-            # it has a default value, that's why we have to jump through the hoops here
-            config_mapping['auto_remove'] = host_config.get('AutoRemove')
-
-        if self.parameters.client.option_minimal_versions['stop_timeout']['supported']:
-            # stop_timeout is only supported in Docker SDK for Python >= 2.1. Note that
-            # stop_timeout has a hybrid role, in that it used to be something only used
-            # for stopping containers, and is now also used as a container property.
-            # That's why it needs special handling here.
-            config_mapping['stop_timeout'] = config.get('StopTimeout')
-
-        differences = DifferenceTracker()
-        for key, value in config_mapping.items():
-            minimal_version = self.parameters.client.option_minimal_versions.get(key, {})
-            if not minimal_version.get('supported', True):
-                continue
-            compare = self.parameters.client.comparisons[self.parameters_map.get(key, key)]
-            self.log('check differences %s %s vs %s (%s)' % (key, getattr(self.parameters, key), to_text(value, errors='surrogate_or_strict'), compare))
-            if getattr(self.parameters, key, None) is not None:
-                match = self._compare(getattr(self.parameters, key), value, compare)
-
-                if not match:
-                    if key == 'expected_healthcheck' and config_mapping['disable_healthcheck'] and self.parameters.disable_healthcheck:
-                        # If the healthcheck is disabled (both in parameters and for the current container), and the user
-                        # requested strict comparison for healthcheck, the comparison will fail. That's why we ignore the
-                        # expected_healthcheck comparison in this case.
-                        continue
-
-                    if key == 'expected_labels' and compare['comparison'] == 'strict' and self.parameters.image_label_mismatch == 'fail':
-                        # If there are labels from the base image that should be removed and
-                        # base_image_mismatch is fail we want raise an error.
-                        image_labels = self._get_image_labels(image)
-                        would_remove_labels = []
-                        for label in image_labels:
-                            if label not in self.parameters.labels:
-                                # Format label for error message
-                                would_remove_labels.append(label)
-                        if would_remove_labels:
-                            msg = ("Some labels should be removed but are present in the base image. You can set image_label_mismatch to 'ignore' to ignore"
-                                   " this error. Labels: {0}")
-                            self.fail(msg.format(', '.join(['"%s"' % label for label in would_remove_labels])))
-
-                    # no match. record the differences
-                    p = getattr(self.parameters, key)
-                    c = value
-                    if compare['type'] == 'set':
-                        # Since the order does not matter, sort so that the diff output is better.
-                        if p is not None:
-                            p = sorted(p)
-                        if c is not None:
-                            c = sorted(c)
-                    elif compare['type'] == 'set(dict)':
-                        # Since the order does not matter, sort so that the diff output is better.
-                        if key == 'expected_mounts':
-                            # For selected values, use one entry as key
-                            def sort_key_fn(x):
-                                return x['target']
-                        else:
-                            # We sort the list of dictionaries by using the sorted items of a dict as its key.
-                            def sort_key_fn(x):
-                                return sorted((a, to_text(b, errors='surrogate_or_strict')) for a, b in x.items())
-                        if p is not None:
-                            p = sorted(p, key=sort_key_fn)
-                        if c is not None:
-                            c = sorted(c, key=sort_key_fn)
-                    differences.add(key, parameter=p, active=c)
-
-        has_differences = not differences.empty
-        return has_differences, differences
-
-    def has_different_resource_limits(self):
-        '''
-        Diff parameters and container resource limits
-        '''
-        if not self.container.get('HostConfig'):
-            self.fail("limits_differ_from_container: Error parsing container properties. HostConfig missing.")
-
-        host_config = self.container['HostConfig']
-
-        restart_policy = host_config.get('RestartPolicy') or dict()
-
-        config_mapping = dict(
-            blkio_weight=host_config.get('BlkioWeight'),
-            cpu_period=host_config.get('CpuPeriod'),
-            cpu_quota=host_config.get('CpuQuota'),
-            cpu_shares=host_config.get('CpuShares'),
-            cpuset_cpus=host_config.get('CpusetCpus'),
-            cpuset_mems=host_config.get('CpusetMems'),
-            kernel_memory=host_config.get("KernelMemory"),
-            memory=host_config.get('Memory'),
-            memory_reservation=host_config.get('MemoryReservation'),
-            memory_swap=host_config.get('MemorySwap'),
-            restart_policy=restart_policy.get('Name')
-        )
-
-        # Options which don't make sense without their accompanying option
-        if self.parameters.restart_policy:
-            config_mapping['restart_retries'] = restart_policy.get('MaximumRetryCount')
-
-        differences = DifferenceTracker()
-        for key, value in config_mapping.items():
-            if getattr(self.parameters, key, None):
-                compare = self.parameters.client.comparisons[self.parameters_map.get(key, key)]
-                match = self._compare(getattr(self.parameters, key), value, compare)
-
-                if not match:
-                    # no match. record the differences
-                    differences.add(key, parameter=getattr(self.parameters, key), active=value)
-        different = not differences.empty
-        return different, differences
-
-    def has_network_differences(self):
-        '''
-        Check if the container is connected to requested networks with expected options: links, aliases, ipv4, ipv6
-        '''
-        different = False
-        differences = []
-
-        if not self.parameters.networks:
-            return different, differences
-
-        if not self.container.get('NetworkSettings'):
-            self.fail("has_missing_networks: Error parsing container properties. NetworkSettings missing.")
-
-        connected_networks = self.container['NetworkSettings']['Networks']
-        for network in self.parameters.networks:
-            network_info = connected_networks.get(network['name'])
-            if network_info is None:
-                different = True
-                differences.append(dict(
-                    parameter=network,
-                    container=None
-                ))
-            else:
-                diff = False
-                network_info_ipam = network_info.get('IPAMConfig') or {}
-                if network.get('ipv4_address') and network['ipv4_address'] != network_info_ipam.get('IPv4Address'):
-                    diff = True
-                if network.get('ipv6_address') and network['ipv6_address'] != network_info_ipam.get('IPv6Address'):
-                    diff = True
-                if network.get('aliases'):
-                    if not compare_generic(network['aliases'], network_info.get('Aliases'), 'allow_more_present', 'set'):
-                        diff = True
-                if network.get('links'):
-                    expected_links = []
-                    for link, alias in network['links']:
-                        expected_links.append("%s:%s" % (link, alias))
-                    if not compare_generic(expected_links, network_info.get('Links'), 'allow_more_present', 'set'):
-                        diff = True
-                if diff:
-                    different = True
-                    differences.append(dict(
-                        parameter=network,
-                        container=dict(
-                            name=network['name'],
-                            ipv4_address=network_info_ipam.get('IPv4Address'),
-                            ipv6_address=network_info_ipam.get('IPv6Address'),
-                            aliases=network_info.get('Aliases'),
-                            links=network_info.get('Links')
-                        )
-                    ))
-        return different, differences
-
-    def has_extra_networks(self):
-        '''
-        Check if the container is connected to non-requested networks
-        '''
-        extra_networks = []
-        extra = False
-
-        if not self.container.get('NetworkSettings'):
-            self.fail("has_extra_networks: Error parsing container properties. NetworkSettings missing.")
-
-        connected_networks = self.container['NetworkSettings'].get('Networks')
-        if connected_networks:
-            for network, network_config in connected_networks.items():
-                keep = False
-                if self.parameters.networks:
-                    for expected_network in self.parameters.networks:
-                        if expected_network['name'] == network:
-                            keep = True
-                if not keep:
-                    extra = True
-                    extra_networks.append(dict(name=network, id=network_config['NetworkID']))
-        return extra, extra_networks
-
-    def _get_expected_devices(self):
-        if not self.parameters.devices:
-            return None
-        expected_devices = []
-        for device in self.parameters.devices:
-            parts = device.split(':')
-            if len(parts) == 1:
-                expected_devices.append(
-                    dict(
-                        CgroupPermissions='rwm',
-                        PathInContainer=parts[0],
-                        PathOnHost=parts[0]
-                    ))
-            elif len(parts) == 2:
-                parts = device.split(':')
-                expected_devices.append(
-                    dict(
-                        CgroupPermissions='rwm',
-                        PathInContainer=parts[1],
-                        PathOnHost=parts[0]
-                    )
-                )
-            else:
-                expected_devices.append(
-                    dict(
-                        CgroupPermissions=parts[2],
-                        PathInContainer=parts[1],
-                        PathOnHost=parts[0]
-                    ))
-        return expected_devices
-
-    def _get_expected_entrypoint(self):
-        if self.parameters.client.module.params['command_handling'] != 'correct' and not self.parameters.entrypoint:
-            return None
-        return self.parameters.entrypoint
-
-    def _get_expected_ports(self):
-        if self.parameters.published_ports is None:
-            return None
-        expected_bound_ports = {}
-        for container_port, config in self.parameters.published_ports.items():
-            if isinstance(container_port, int):
-                container_port = "%s/tcp" % container_port
-            if len(config) == 1:
-                if isinstance(config[0], int):
-                    expected_bound_ports[container_port] = [{'HostIp': "0.0.0.0", 'HostPort': config[0]}]
-                else:
-                    expected_bound_ports[container_port] = [{'HostIp': config[0], 'HostPort': ""}]
-            elif isinstance(config[0], tuple):
-                expected_bound_ports[container_port] = []
-                for host_ip, host_port in config:
-                    expected_bound_ports[container_port].append({'HostIp': host_ip, 'HostPort': to_text(host_port, errors='surrogate_or_strict')})
-            else:
-                expected_bound_ports[container_port] = [{'HostIp': config[0], 'HostPort': to_text(config[1], errors='surrogate_or_strict')}]
-        return expected_bound_ports
-
-    def _get_expected_links(self):
-        if self.parameters.links is None:
-            return None
-        self.log('parameter links:')
-        self.log(self.parameters.links, pretty_print=True)
-        exp_links = []
-        for link, alias in self.parameters.links:
-            exp_links.append("/%s:%s/%s" % (link, ('/' + self.parameters.name), alias))
-        return exp_links
-
-    def _get_expected_binds(self, image):
-        self.log('_get_expected_binds')
-        image_vols = []
-        if image:
-            image_vols = self._get_image_binds(image['Config'].get('Volumes'))
-        param_vols = []
-        if self.parameters.volumes:
-            for vol in self.parameters.volumes:
-                host = None
-                if ':' in vol:
-                    parts = vol.split(':')
-                    if len(parts) == 3:
-                        host, container, mode = parts
-                        if not is_volume_permissions(mode):
-                            self.fail('Found invalid volumes mode: {0}'.format(mode))
-                    if len(parts) == 2:
-                        if not is_volume_permissions(parts[1]):
-                            host, container, mode = parts + ['rw']
-                if host:
-                    param_vols.append("%s:%s:%s" % (host, container, mode))
-        result = list(set(image_vols + param_vols))
-        self.log("expected_binds:")
-        self.log(result, pretty_print=True)
-        return result
-
-    def _get_expected_labels(self, image):
-        if self.parameters.labels is None:
-            return None
-        if self.parameters.image_label_mismatch == 'ignore':
-            expected_labels = dict(self._get_image_labels(image))
-        else:
-            expected_labels = {}
-        expected_labels.update(self.parameters.labels)
-        return expected_labels
-
-    def _get_image_labels(self, image):
-        if not image:
-            return {}
-
-        # Can't use get('Labels', {}) because 'Labels' may be present and be None
-        return image['Config'].get('Labels') or {}
-
-    def _get_expected_device_requests(self):
-        if self.parameters.device_requests is None:
-            return None
-        device_requests = []
-        for dr in self.parameters.device_requests:
-            device_requests.append({
-                'Driver': dr['driver'],
-                'Count': dr['count'],
-                'DeviceIDs': dr['device_ids'],
-                'Capabilities': dr['capabilities'],
-                'Options': dr['options'],
-            })
-        return device_requests
-
-    def _get_image_binds(self, volumes):
-        '''
-        Convert array of binds to array of strings with format host_path:container_path:mode
-
-        :param volumes: array of bind dicts
-        :return: array of strings
-        '''
-        results = []
-        if isinstance(volumes, dict):
-            results += self._get_bind_from_dict(volumes)
-        elif isinstance(volumes, list):
-            for vol in volumes:
-                results += self._get_bind_from_dict(vol)
-        return results
-
-    @staticmethod
-    def _get_bind_from_dict(volume_dict):
-        results = []
-        if volume_dict:
-            for host_path, config in volume_dict.items():
-                if isinstance(config, dict) and config.get('bind'):
-                    container_path = config.get('bind')
-                    mode = config.get('mode', 'rw')
-                    results.append("%s:%s:%s" % (host_path, container_path, mode))
-        return results
-
-    def _get_expected_volumes(self, image):
-        self.log('_get_expected_volumes')
-        expected_vols = dict()
-        if image and image['Config'].get('Volumes'):
-            expected_vols.update(image['Config'].get('Volumes'))
-
-        if self.parameters.volumes:
-            for vol in self.parameters.volumes:
-                # We only expect anonymous volumes to show up in the list
-                if ':' in vol:
-                    parts = vol.split(':')
-                    if len(parts) == 3:
-                        continue
-                    if len(parts) == 2:
-                        if not is_volume_permissions(parts[1]):
-                            continue
-                expected_vols[vol] = dict()
-
-        if not expected_vols:
-            expected_vols = None
-        self.log("expected_volumes:")
-        self.log(expected_vols, pretty_print=True)
-        return expected_vols
-
-    def _get_expected_env(self, image):
-        self.log('_get_expected_env')
-        expected_env = dict()
-        if image and image['Config'].get('Env'):
-            for env_var in image['Config']['Env']:
-                parts = env_var.split('=', 1)
-                expected_env[parts[0]] = parts[1]
-        if self.parameters.env:
-            expected_env.update(self.parameters.env)
-        param_env = []
-        for key, value in expected_env.items():
-            param_env.append("%s=%s" % (key, value))
-        return param_env
-
-    def _get_expected_exposed(self, image):
-        self.log('_get_expected_exposed')
-        image_ports = []
-        if image:
-            image_exposed_ports = image['Config'].get('ExposedPorts') or {}
-            image_ports = [self._normalize_port(p) for p in image_exposed_ports.keys()]
-        param_ports = []
-        if self.parameters.ports:
-            param_ports = [to_text(p[0], errors='surrogate_or_strict') + '/' + p[1] for p in self.parameters.ports]
-        result = list(set(image_ports + param_ports))
-        self.log(result, pretty_print=True)
-        return result
-
-    def _get_expected_ulimits(self, config_ulimits):
-        self.log('_get_expected_ulimits')
-        if config_ulimits is None:
-            return None
-        results = []
-        for limit in config_ulimits:
-            results.append(dict(
-                Name=limit.name,
-                Soft=limit.soft,
-                Hard=limit.hard
-            ))
-        return results
-
-    def _get_expected_sysctls(self, config_sysctls):
-        self.log('_get_expected_sysctls')
-        if config_sysctls is None:
-            return None
-        result = dict()
-        for key, value in config_sysctls.items():
-            result[key] = to_text(value, errors='surrogate_or_strict')
-        return result
-
-    def _get_expected_cmd(self):
-        self.log('_get_expected_cmd')
-        if self.parameters.client.module.params['command_handling'] != 'correct' and not self.parameters.command:
-            return None
-        return self.parameters.command
-
-    def _convert_simple_dict_to_list(self, param_name, join_with=':'):
-        if getattr(self.parameters, param_name, None) is None:
-            return None
-        results = []
-        for key, value in getattr(self.parameters, param_name).items():
-            results.append("%s%s%s" % (key, join_with, value))
-        return results
-
-    def _normalize_port(self, port):
-        if '/' not in port:
-            return port + '/tcp'
-        return port
-
-    def _get_expected_healthcheck(self):
-        self.log('_get_expected_healthcheck')
-        expected_healthcheck = dict()
-
-        if self.parameters.healthcheck:
-            expected_healthcheck.update([(k.title().replace("_", ""), v)
-                                         for k, v in self.parameters.healthcheck.items()])
-
-        return expected_healthcheck
-
 
 class ContainerManager(DockerBaseClass):
-    '''
-    Perform container management tasks
-    '''
-
-    def __init__(self, client):
-
-        super(ContainerManager, self).__init__()
-
-        if client.module.params.get('log_options') and not client.module.params.get('log_driver'):
-            client.module.warn('log_options is ignored when log_driver is not specified')
-        if client.module.params.get('healthcheck') and not client.module.params.get('healthcheck').get('test'):
-            client.module.warn('healthcheck is ignored when test is not specified')
-        if client.module.params.get('restart_retries') is not None and not client.module.params.get('restart_policy'):
-            client.module.warn('restart_retries is ignored when restart_policy is not specified')
-
+    def __init__(self, module, client, active_options):
         self.client = client
-        self.parameters = TaskParameters(client)
-        self.check_mode = self.client.check_mode
+        self.options = active_options
+        self.all_options = self._collect_all_options(active_options)
+        self.module = module
+        self.check_mode = self.module.check_mode
+        self.param_cleanup = self.module.params['cleanup']
+        self.param_container_default_behavior = self.module.params['container_default_behavior']
+        self.param_default_host_ip = self.module.params['default_host_ip']
+        self.param_debug = self.module.params['debug']
+        self.param_force_kill = self.module.params['force_kill']
+        self.param_image = self.module.params['image']
+        self.param_image_label_mismatch = self.module.params['image_label_mismatch']
+        self.param_keep_volumes = self.module.params['keep_volumes']
+        self.param_kill_signal = self.module.params['kill_signal']
+        self.param_name = self.module.params['name']
+        self.param_networks_cli_compatible = self.module.params['networks_cli_compatible']
+        self.param_output_logs = self.module.params['output_logs']
+        self.param_paused = self.module.params['paused']
+        self.param_pull = self.module.params['pull']
+        self.param_purge_networks = self.module.params['purge_networks']
+        self.param_recreate = self.module.params['recreate']
+        self.param_removal_wait_timeout = self.module.params['removal_wait_timeout']
+        self.param_restart = self.module.params['restart']
+        self.param_state = self.module.params['state']
+        self._parse_comparisons()
+        self._update_params()
+        self.parameters = self._collect_params(active_options)
         self.results = {'changed': False, 'actions': []}
         self.diff = {}
         self.diff_tracker = DifferenceTracker()
         self.facts = {}
+        if self.param_default_host_ip:
+            valid_ip = False
+            if re.match(r'^[0-9]+\.[0-9]+\.[0-9]+\.[0-9]+$', self.param_default_host_ip):
+                valid_ip = True
+            if re.match(r'^\[[0-9a-fA-F:]+\]$', self.param_default_host_ip):
+                valid_ip = True
+            if re.match(r'^[0-9a-fA-F:]+$', self.param_default_host_ip):
+                self.param_default_host_ip = '[{0}]'.format(self.param_default_host_ip)
+                valid_ip = True
+            if not valid_ip:
+                self.fail('The value of default_host_ip must be an empty string, an IPv4 address, '
+                          'or an IPv6 address. Got "{0}" instead.'.format(self.param_default_host_ip))
 
-        state = self.parameters.state
-        if state in ('stopped', 'started', 'present'):
-            self.present(state)
-        elif state == 'absent':
+
+    def _collect_all_options(self, active_options):
+        all_options = {}
+        for options in active_options:
+            for option in options.options:
+                all_options[option.name] = option
+        for option in [
+            Option('image', 'str', None),
+            Option('networks', 'set', None, elements='dict', ansible_suboptions={}),
+        ]:
+            all_options[option.name] = option
+        return all_options
+
+    def _collect_all_module_params(self):
+        all_module_options = set()
+        for option, data in self.module.argument_spec.items():
+            all_module_options.add(option)
+            if 'aliases' in data:
+                for alias in data['aliases']:
+                    all_module_options.add(alias)
+        return all_module_options
+
+    def _parse_comparisons(self):
+        # Keep track of all module params and all option aliases
+        all_module_options = self._collect_all_module_params()
+        comp_aliases = {}
+        for option_name, option in self.all_options.items():
+            if option.not_an_ansible_option:
+                continue
+            comp_aliases[option_name] = option_name
+            for alias in option.ansible_aliases:
+                comp_aliases[alias] = option_name
+        # Process legacy ignore options
+        if self.module.params['ignore_image']:
+            self.all_options['image'].comparison = 'ignore'
+        if self.param_purge_networks:
+            self.all_options['networks'].comparison = 'strict'
+        # Process comparsions specified by user
+        if self.module.params.get('comparisons'):
+            # If '*' appears in comparisons, process it first
+            if '*' in self.module.params['comparisons']:
+                value = self.module.params['comparisons']['*']
+                if value not in ('strict', 'ignore'):
+                    self.fail("The wildcard can only be used with comparison modes 'strict' and 'ignore'!")
+                for option in self.all_options.values():
+                    if option.name == 'networks':
+                        # `networks` is special: only update if
+                        # some value is actually specified
+                        if self.module.params['networks'] is None:
+                            continue
+                    option.comparison = value
+            # Now process all other comparisons.
+            comp_aliases_used = {}
+            for key, value in self.module.params['comparisons'].items():
+                if key == '*':
+                    continue
+                # Find main key
+                key_main = comp_aliases.get(key)
+                if key_main is None:
+                    if key_main in all_module_options:
+                        self.fail("The module option '%s' cannot be specified in the comparisons dict, "
+                                  "since it does not correspond to container's state!" % key)
+                    if key not in self.all_options or self.all_options[key].not_an_ansible_option:
+                        self.fail("Unknown module option '%s' in comparisons dict!" % key)
+                    key_main = key
+                if key_main in comp_aliases_used:
+                    self.fail("Both '%s' and '%s' (aliases of %s) are specified in comparisons dict!" % (key, comp_aliases_used[key_main], key_main))
+                comp_aliases_used[key_main] = key
+                # Check value and update accordingly
+                if value in ('strict', 'ignore'):
+                    self.all_options[key_main].comparison = value
+                elif value == 'allow_more_present':
+                    if self.all_options[key_main].comparison_type == 'value':
+                        self.fail("Option '%s' is a value and not a set/list/dict, so its comparison cannot be %s" % (key, value))
+                    self.all_options[key_main].comparison = value
+                else:
+                    self.fail("Unknown comparison mode '%s'!" % value)
+        # Copy values
+        for option in self.all_options.values():
+            if option.copy_comparison_from is not None:
+                option.comparison = self.all_options[option.copy_comparison_from].comparison
+        # Check legacy values
+        if self.module.params['ignore_image'] and self.all_options['image'].comparison != 'ignore':
+            self.module.warn('The ignore_image option has been overridden by the comparisons option!')
+        if self.param_purge_networks and self.all_options['networks'].comparison != 'strict':
+            self.module.warn('The purge_networks option has been overridden by the comparisons option!')
+
+    def _update_params(self):
+        if self.param_networks_cli_compatible is True and self.module.params['networks'] and self.module.params['network_mode'] is None:
+            # Same behavior as Docker CLI: if networks are specified, use the name of the first network as the value for network_mode
+            # (assuming no explicit value is specified for network_mode)
+            self.module.params['network_mode'] = self.module.params['networks'][0]['name']
+        if self.param_container_default_behavior == 'compatibility':
+            old_default_values = dict(
+                auto_remove=False,
+                detach=True,
+                init=False,
+                interactive=False,
+                memory='0',
+                paused=False,
+                privileged=False,
+                read_only=False,
+                tty=False,
+            )
+            for param, value in old_default_values.items():
+                if self.module.params[param] is None:
+                    self.module.params[param] = value
+
+    def _collect_params(self, active_options):
+        parameters = []
+        for options in active_options:
+            values = {}
+            engine = options.get_engine('docker_api')
+            for option in options.options:
+                if not option.not_an_ansible_option and self.module.params[option.name] is not None:
+                    values[option.name] = self.module.params[option.name]
+            values = options.preprocess(self.module, values)
+            engine.preprocess_value(self.module, self.client, self.client.docker_api_version, options.options, values)
+            parameters.append((options, values))
+        return parameters
+
+    def fail(self, *args, **kwargs):
+        self.client.fail(*args, **kwargs)
+
+    def run(self):
+        if self.param_state in ('stopped', 'started', 'present'):
+            self.present(self.param_state)
+        elif self.param_state == 'absent':
             self.absent()
 
-        if not self.check_mode and not self.parameters.debug:
+        if not self.check_mode and not self.param_debug:
             self.results.pop('actions')
 
-        if self.client.module._diff or self.parameters.debug:
+        if self.module._diff or self.param_debug:
             self.diff['before'], self.diff['after'] = self.diff_tracker.get_before_after()
             self.results['diff'] = self.diff
 
@@ -2852,7 +1493,7 @@ class ContainerManager(DockerBaseClass):
             delay = min(delay * 1.1, 10)
 
     def present(self, state):
-        container = self._get_container(self.parameters.name)
+        container = self._get_container(self.param_name)
         was_running = container.running
         was_paused = container.paused
         container_created = False
@@ -2870,31 +1511,31 @@ class ContainerManager(DockerBaseClass):
                 self.log('Found container in removal phase')
             else:
                 self.log('No container found')
-            if not self.parameters.image:
+            if not self.param_image:
                 self.fail('Cannot create container when image is not specified!')
             self.diff_tracker.add('exists', parameter=True, active=False)
             if container.removing and not self.check_mode:
                 # Wait for container to be removed before trying to create it
                 self.wait_for_state(
-                    container.Id, wait_states=['removing'], accept_removal=True, max_wait=self.parameters.removal_wait_timeout)
-            new_container = self.container_create(self.parameters.image, self.parameters.create_parameters)
+                    container.Id, wait_states=['removing'], accept_removal=True, max_wait=self.param_removal_wait_timeout)
+            new_container = self.container_create(self.param_image)
             if new_container:
                 container = new_container
             container_created = True
         else:
             # Existing container
-            different, differences = container.has_different_configuration(image)
+            different, differences = self.has_different_configuration(container, image)
             image_different = False
-            if self.parameters.comparisons['image']['comparison'] == 'strict':
+            if self.all_options['image'].comparison == 'strict':
                 image_different = self._image_is_different(image, container)
-            if image_different or different or self.parameters.recreate:
+            if image_different or different or self.param_recreate:
                 self.diff_tracker.merge(differences)
                 self.diff['differences'] = differences.get_legacy_docker_container_diffs()
                 if image_different:
                     self.diff['image_different'] = True
                 self.log("differences")
                 self.log(differences.get_legacy_docker_container_diffs(), pretty_print=True)
-                image_to_use = self.parameters.image
+                image_to_use = self.param_image
                 if not image_to_use and container and container.Image:
                     image_to_use = container.Image
                 if not image_to_use:
@@ -2904,20 +1545,20 @@ class ContainerManager(DockerBaseClass):
                 self.container_remove(container.Id)
                 if not self.check_mode:
                     self.wait_for_state(
-                        container.Id, wait_states=['removing'], accept_removal=True, max_wait=self.parameters.removal_wait_timeout)
-                new_container = self.container_create(image_to_use, self.parameters.create_parameters)
+                        container.Id, wait_states=['removing'], accept_removal=True, max_wait=self.param_removal_wait_timeout)
+                new_container = self.container_create(image_to_use)
                 if new_container:
                     container = new_container
                 container_created = True
 
         if container and container.exists:
-            container = self.update_limits(container)
+            container = self.update_limits(container, image)
             container = self.update_networks(container, container_created)
 
             if state == 'started' and not container.running:
                 self.diff_tracker.add('running', parameter=True, active=was_running)
                 container = self.container_start(container.Id)
-            elif state == 'started' and self.parameters.restart:
+            elif state == 'started' and self.param_restart:
                 self.diff_tracker.add('running', parameter=True, active=was_running)
                 self.diff_tracker.add('restarted', parameter=True, active=False)
                 container = self.container_restart(container.Id)
@@ -2926,26 +1567,26 @@ class ContainerManager(DockerBaseClass):
                 self.container_stop(container.Id)
                 container = self._get_container(container.Id)
 
-            if state == 'started' and self.parameters.paused is not None and container.paused != self.parameters.paused:
-                self.diff_tracker.add('paused', parameter=self.parameters.paused, active=was_paused)
+            if state == 'started' and self.param_paused is not None and container.paused != self.param_paused:
+                self.diff_tracker.add('paused', parameter=self.param_paused, active=was_paused)
                 if not self.check_mode:
                     try:
-                        if self.parameters.paused:
-                            self.client.pause(container=container.Id)
+                        if self.param_paused:
+                            self.client.post_call('/containers/{0}/pause', container.Id)
                         else:
-                            self.client.unpause(container=container.Id)
+                            self.client.post_call('/containers/{0}/unpause', container.Id)
                     except Exception as exc:
                         self.fail("Error %s container %s: %s" % (
-                            "pausing" if self.parameters.paused else "unpausing", container.Id, to_native(exc)
+                            "pausing" if self.param_paused else "unpausing", container.Id, to_native(exc)
                         ))
                     container = self._get_container(container.Id)
                 self.results['changed'] = True
-                self.results['actions'].append(dict(set_paused=self.parameters.paused))
+                self.results['actions'].append(dict(set_paused=self.param_paused))
 
         self.facts = container.raw
 
     def absent(self):
-        container = self._get_container(self.parameters.name)
+        container = self._get_container(self.param_name)
         if container.exists:
             if container.running:
                 self.diff_tracker.add('running', parameter=False, active=True)
@@ -2953,30 +1594,28 @@ class ContainerManager(DockerBaseClass):
             self.diff_tracker.add('exists', parameter=False, active=True)
             self.container_remove(container.Id)
 
-    def fail(self, msg, **kwargs):
-        self.client.fail(msg, **kwargs)
-
     def _output_logs(self, msg):
-        self.client.module.log(msg=msg)
+        self.module.log(msg=msg)
 
     def _get_container(self, container):
         '''
         Expects container ID or Name. Returns a container object
         '''
-        return Container(self.client.get_container(container), self.parameters)
+        return Container(self.client.get_container(container))
 
     def _get_image(self):
-        if not self.parameters.image:
+        image_parameter = self.param_image
+        if not image_parameter:
             self.log('No image specified')
             return None
-        if is_image_name_id(self.parameters.image):
-            image = self.client.find_image_by_id(self.parameters.image)
+        if is_image_name_id(image_parameter):
+            image = self.client.find_image_by_id(image_parameter)
         else:
-            repository, tag = utils.parse_repository_tag(self.parameters.image)
+            repository, tag = parse_repository_tag(image_parameter)
             if not tag:
                 tag = "latest"
             image = self.client.find_image(repository, tag)
-            if not image or self.parameters.pull:
+            if not image or self.param_pull:
                 if not self.check_mode:
                     self.log("Pull the image.")
                     image, alreadyToLatest = self.client.pull_image(repository, tag)
@@ -3003,21 +1642,188 @@ class ContainerManager(DockerBaseClass):
                     return True
         return False
 
-    def update_limits(self, container):
-        limits_differ, different_limits = container.has_different_resource_limits()
+    def _compose_create_parameters(self, image):
+        params = {
+            'Image': image,
+        }
+        for options, values in self.parameters:
+            engine = options.get_engine('docker_api')
+            if engine.can_set_value(self.client.docker_api_version):
+                engine.set_value(self.module, params, self.client.docker_api_version, options.options, values)
+        return params
+
+    def has_different_configuration(self, container, image):
+        differences = DifferenceTracker()
+        for options, param_values in self.parameters:
+            engine = options.get_engine('docker_api')
+            container_values = engine.get_value(self.module, container.raw, self.client.docker_api_version, options.options)
+            expected_values = engine.get_expected_values(self.module, self.client, self.client.docker_api_version, options.options, image, param_values.copy())
+            for option in options.options:
+                if option.name in expected_values:
+                    param_value = expected_values[option.name]
+                    container_value = container_values.get(option.name)
+                    match = compare_generic(param_value, container_value, option.comparison, option.comparison_type)
+
+                    if not match:
+                        if engine.ignore_mismatching_result(self.module, self.client, self.client.docker_api_version, option, image, container_value, param_value):
+                            continue
+                        # TODO
+                        # if option.name == 'healthcheck' and config_mapping['disable_healthcheck'] and self.parameters.disable_healthcheck:
+                        #     # If the healthcheck is disabled (both in parameters and for the current container), and the user
+                        #     # requested strict comparison for healthcheck, the comparison will fail. That's why we ignore the
+                        #     # expected_healthcheck comparison in this case.
+                        #     continue
+
+                        # no match. record the differences
+                        p = param_value
+                        c = container_value
+                        if option.comparison_type == 'set':
+                            # Since the order does not matter, sort so that the diff output is better.
+                            if p is not None:
+                                p = sorted(p)
+                            if c is not None:
+                                c = sorted(c)
+                        elif option.comparison_type == 'set(dict)':
+                            # Since the order does not matter, sort so that the diff output is better.
+                            if option.name == 'expected_mounts':
+                                # For selected values, use one entry as key
+                                def sort_key_fn(x):
+                                    return x['target']
+                            else:
+                                # We sort the list of dictionaries by using the sorted items of a dict as its key.
+                                def sort_key_fn(x):
+                                    return sorted((a, to_text(b, errors='surrogate_or_strict')) for a, b in x.items())
+                            if p is not None:
+                                p = sorted(p, key=sort_key_fn)
+                            if c is not None:
+                                c = sorted(c, key=sort_key_fn)
+                        differences.add(option.name, parameter=p, active=c)
+
+        has_differences = not differences.empty
+        return has_differences, differences
+
+    def has_different_resource_limits(self, container, image):
+        '''
+        Diff parameters and container resource limits
+        '''
+        differences = DifferenceTracker()
+        for options, param_values in self.parameters:
+            engine = options.get_engine('docker_api')
+            if not engine.can_update_value(self.client.docker_api_version):
+                continue
+            container_values = engine.get_value(self.module, container.raw, self.client.docker_api_version, options.options)
+            expected_values = engine.get_expected_values(self.module, self.client, self.client.docker_api_version, options.options, image, param_values.copy())
+            for option in options.options:
+                if option.name in expected_values:
+                    param_value = expected_values[option.name]
+                    container_value = container_values.get(option.name)
+                    match = compare_generic(param_value, container_value, option.comparison, option.comparison_type)
+
+                    if not match:
+                        # no match. record the differences
+                        differences.add(option.name, parameter=param_value, active=container_value)
+        different = not differences.empty
+        return different, differences
+
+    def _compose_update_parameters(self):
+        result = {}
+        for options, values in self.parameters:
+            engine = options.get_engine('docker_api')
+            if not engine.can_update_value(self.client.docker_api_version):
+                continue
+            engine.update_value(self.module, result, self.client.docker_api_version, options.options, values)
+        return result
+
+    def update_limits(self, container, image):
+        limits_differ, different_limits = self.has_different_resource_limits(container, image)
         if limits_differ:
             self.log("limit differences:")
             self.log(different_limits.get_legacy_docker_container_diffs(), pretty_print=True)
             self.diff_tracker.merge(different_limits)
         if limits_differ and not self.check_mode:
-            self.container_update(container.Id, self.parameters.update_parameters)
+            self.container_update(container.Id, self._compose_update_parameters())
             return self._get_container(container.Id)
         return container
 
+    def has_network_differences(self, container):
+        '''
+        Check if the container is connected to requested networks with expected options: links, aliases, ipv4, ipv6
+        '''
+        different = False
+        differences = []
+
+        if not self.module.params['networks']:
+            return different, differences
+
+        if not container.container.get('NetworkSettings'):
+            self.fail("has_missing_networks: Error parsing container properties. NetworkSettings missing.")
+
+        connected_networks = container.container['NetworkSettings']['Networks']
+        for network in self.module.params['networks']:
+            network_info = connected_networks.get(network['name'])
+            if network_info is None:
+                different = True
+                differences.append(dict(
+                    parameter=network,
+                    container=None
+                ))
+            else:
+                diff = False
+                network_info_ipam = network_info.get('IPAMConfig') or {}
+                if network.get('ipv4_address') and network['ipv4_address'] != network_info_ipam.get('IPv4Address'):
+                    diff = True
+                if network.get('ipv6_address') and network['ipv6_address'] != network_info_ipam.get('IPv6Address'):
+                    diff = True
+                if network.get('aliases'):
+                    if not compare_generic(network['aliases'], network_info.get('Aliases'), 'allow_more_present', 'set'):
+                        diff = True
+                if network.get('links'):
+                    expected_links = []
+                    for link, alias in network['links']:
+                        expected_links.append("%s:%s" % (link, alias))
+                    if not compare_generic(expected_links, network_info.get('Links'), 'allow_more_present', 'set'):
+                        diff = True
+                if diff:
+                    different = True
+                    differences.append(dict(
+                        parameter=network,
+                        container=dict(
+                            name=network['name'],
+                            ipv4_address=network_info_ipam.get('IPv4Address'),
+                            ipv6_address=network_info_ipam.get('IPv6Address'),
+                            aliases=network_info.get('Aliases'),
+                            links=network_info.get('Links')
+                        )
+                    ))
+        return different, differences
+
+    def has_extra_networks(self, container):
+        '''
+        Check if the container is connected to non-requested networks
+        '''
+        extra_networks = []
+        extra = False
+
+        if not container.container.get('NetworkSettings'):
+            self.fail("has_extra_networks: Error parsing container properties. NetworkSettings missing.")
+
+        connected_networks = container.container['NetworkSettings'].get('Networks')
+        if connected_networks:
+            for network, network_config in connected_networks.items():
+                keep = False
+                if self.module.params['networks']:
+                    for expected_network in self.module.params['networks']:
+                        if expected_network['name'] == network:
+                            keep = True
+                if not keep:
+                    extra = True
+                    extra_networks.append(dict(name=network, id=network_config['NetworkID']))
+        return extra, extra_networks
+
     def update_networks(self, container, container_created):
         updated_container = container
-        if self.parameters.comparisons['networks']['comparison'] != 'ignore' or container_created:
-            has_network_differences, network_differences = container.has_network_differences()
+        if self.all_options['networks'].comparison != 'ignore' or container_created:
+            has_network_differences, network_differences = self.has_network_differences(container)
             if has_network_differences:
                 if self.diff.get('differences'):
                     self.diff['differences'].append(dict(network_differences=network_differences))
@@ -3032,8 +1838,8 @@ class ContainerManager(DockerBaseClass):
                 self.results['changed'] = True
                 updated_container = self._add_networks(container, network_differences)
 
-        if (self.parameters.comparisons['networks']['comparison'] == 'strict' and self.parameters.networks is not None) or self.parameters.purge_networks:
-            has_extra_networks, extra_networks = container.has_extra_networks()
+        if (self.all_options['networks'].comparison == 'strict' and self.module.params['networks'] is not None) or self.param_purge_networks:
+            has_extra_networks, extra_networks = self.has_extra_networks(container)
             if has_extra_networks:
                 if self.diff.get('differences'):
                     self.diff['differences'].append(dict(purge_networks=extra_networks))
@@ -3055,21 +1861,28 @@ class ContainerManager(DockerBaseClass):
                 self.results['actions'].append(dict(removed_from_network=diff['parameter']['name']))
                 if not self.check_mode:
                     try:
-                        self.client.disconnect_container_from_network(container.Id, diff['parameter']['id'])
+                        self.client.post_json('/networks/{0}/disconnect', diff['parameter']['id'], data={'Container': container.Id})
                     except Exception as exc:
                         self.fail("Error disconnecting container from network %s - %s" % (diff['parameter']['name'],
                                                                                           to_native(exc)))
             # connect to the network
             params = dict()
-            for para in ('ipv4_address', 'ipv6_address', 'links', 'aliases'):
+            for para, dest_para in {'ipv4_address': 'IPv4Address', 'ipv6_address': 'IPv6Address', 'links': 'Links', 'aliases': 'Aliases'}.items():
                 if diff['parameter'].get(para):
-                    params[para] = diff['parameter'][para]
+                    value = diff['parameter'][para]
+                    if para == 'links':
+                        value = normalize_links(value)
+                    params[dest_para] = value
             self.results['actions'].append(dict(added_to_network=diff['parameter']['name'], network_parameters=params))
             if not self.check_mode:
                 try:
                     self.log("Connecting container to network %s" % diff['parameter']['id'])
                     self.log(params, pretty_print=True)
-                    self.client.connect_container_to_network(container.Id, diff['parameter']['id'], **params)
+                    data = {
+                        'Container': container.Id,
+                        'EndpointConfig': params,
+                    }
+                    self.client.post_json('/networks/{0}/connect', diff['parameter']['id'], data=data)
                 except Exception as exc:
                     self.fail("Error connecting container to network %s - %s" % (diff['parameter']['name'], to_native(exc)))
         return self._get_container(container.Id)
@@ -3079,13 +1892,14 @@ class ContainerManager(DockerBaseClass):
             self.results['actions'].append(dict(removed_from_network=network['name']))
             if not self.check_mode:
                 try:
-                    self.client.disconnect_container_from_network(container.Id, network['name'])
+                    self.client.post_json('/networks/{0}/disconnect', network['name'], data={'Container': container.Id})
                 except Exception as exc:
                     self.fail("Error disconnecting container from network %s - %s" % (network['name'],
                                                                                       to_native(exc)))
         return self._get_container(container.Id)
 
-    def container_create(self, image, create_parameters):
+    def container_create(self, image):
+        create_parameters = self._compose_create_parameters(image)
         self.log("create container")
         self.log("image: %s parameters:" % image)
         self.log(create_parameters, pretty_print=True)
@@ -3094,7 +1908,8 @@ class ContainerManager(DockerBaseClass):
         new_container = None
         if not self.check_mode:
             try:
-                new_container = self.client.create_container(image, **create_parameters)
+                params = {'name': self.param_name}
+                new_container = self.client.post_json_to_json('/containers/create', data=create_parameters, params=params)
                 self.client.report_warnings(new_container)
             except Exception as exc:
                 self.fail("Error creating container: %s" % to_native(exc))
@@ -3107,34 +1922,39 @@ class ContainerManager(DockerBaseClass):
         self.results['changed'] = True
         if not self.check_mode:
             try:
-                self.client.start(container=container_id)
+                self.client.post_json('/containers/{0}/start', container_id)
             except Exception as exc:
                 self.fail("Error starting container %s: %s" % (container_id, to_native(exc)))
 
-            if self.parameters.detach is False:
-                if self.client.docker_py_version >= LooseVersion('3.0'):
-                    status = self.client.wait(container_id)['StatusCode']
-                else:
-                    status = self.client.wait(container_id)
+            if self.module.params['detach'] is False:
+                status = self.client.post_json_as_json('/containers/{0}/wait', container_id)['StatusCode']
                 self.client.fail_results['status'] = status
                 self.results['status'] = status
 
-                if self.parameters.auto_remove:
+                if self.module.params['auto_remove']:
                     output = "Cannot retrieve result as auto_remove is enabled"
-                    if self.parameters.output_logs:
-                        self.client.module.warn('Cannot output_logs if auto_remove is enabled!')
+                    if self.param_output_logs:
+                        self.module.warn('Cannot output_logs if auto_remove is enabled!')
                 else:
-                    config = self.client.inspect_container(container_id)
+                    config = self.client.get_json('/containers/{0}/json', container_id)
                     logging_driver = config['HostConfig']['LogConfig']['Type']
 
                     if logging_driver in ('json-file', 'journald', 'local'):
-                        output = self.client.logs(container_id, stdout=True, stderr=True, stream=False, timestamps=False)
-                        if self.parameters.output_logs:
+                        params = {
+                            'stderr': 1,
+                            'stdout': 1,
+                            'timestamps': 0,
+                            'follow': 0,
+                            'tail': 'all',
+                        }
+                        res = self.client._get(self.client._url('/containers/{0}/logs', container_id), params=params)
+                        output = self.client._get_result_tty(False, res, config['Config']['Tty'])
+                        if self.param_output_logs:
                             self._output_logs(msg=output)
                     else:
                         output = "Result logged using `%s` driver" % logging_driver
 
-                if self.parameters.cleanup:
+                if self.param_cleanup:
                     self.container_remove(container_id, force=True)
                 insp = self._get_container(container_id)
                 if insp.raw:
@@ -3149,16 +1969,16 @@ class ContainerManager(DockerBaseClass):
         return self._get_container(container_id)
 
     def container_remove(self, container_id, link=False, force=False):
-        volume_state = (not self.parameters.keep_volumes)
+        volume_state = (not self.param_keep_volumes)
         self.log("remove container container:%s v:%s link:%s force%s" % (container_id, volume_state, link, force))
         self.results['actions'].append(dict(removed=container_id, volume_state=volume_state, link=link, force=force))
         self.results['changed'] = True
-        response = None
         if not self.check_mode:
             count = 0
             while True:
                 try:
-                    response = self.client.remove_container(container_id, v=volume_state, link=link, force=force)
+                    params = {'v': volume_state, 'link': link, 'force': force}
+                    self.client.delete_call('/containers/{0}', container_id, params=params)
                 except NotFound as dummy:
                     pass
                 except APIError as exc:
@@ -3170,7 +1990,7 @@ class ContainerManager(DockerBaseClass):
                         count += 1
                         # Unpause
                         try:
-                            self.client.unpause(container=container_id)
+                            self.client.post_call('/containers/{0}/unpause', container_id)
                         except Exception as exc2:
                             self.fail("Error unpausing container %s for removal: %s" % (container_id, to_native(exc2)))
                         # Now try again
@@ -3183,7 +2003,6 @@ class ContainerManager(DockerBaseClass):
                     self.fail("Error removing container %s: %s" % (container_id, to_native(exc)))
                 # We only loop when explicitly requested by 'continue'
                 break
-        return response
 
     def container_update(self, container_id, update_parameters):
         if update_parameters:
@@ -3193,54 +2012,58 @@ class ContainerManager(DockerBaseClass):
             self.results['changed'] = True
             if not self.check_mode and callable(getattr(self.client, 'update_container')):
                 try:
-                    result = self.client.update_container(container_id, **update_parameters)
+                    result = self.client.post_json_to_json('/containers/{0}/update', container_id, data=update_parameters)
                     self.client.report_warnings(result)
                 except Exception as exc:
                     self.fail("Error updating container %s: %s" % (container_id, to_native(exc)))
         return self._get_container(container_id)
 
     def container_kill(self, container_id):
-        self.results['actions'].append(dict(killed=container_id, signal=self.parameters.kill_signal))
+        self.results['actions'].append(dict(killed=container_id, signal=self.param_kill_signal))
         self.results['changed'] = True
-        response = None
         if not self.check_mode:
             try:
-                if self.parameters.kill_signal:
-                    response = self.client.kill(container_id, signal=self.parameters.kill_signal)
-                else:
-                    response = self.client.kill(container_id)
+                params = {}
+                if self.param_kill_signal is not None:
+                    params['signal'] = int(self.param_kill_signal)
+                self.client.post_call('/containers/{0}/kill', container_id, params=params)
             except Exception as exc:
                 self.fail("Error killing container %s: %s" % (container_id, to_native(exc)))
-        return response
 
     def container_restart(self, container_id):
-        self.results['actions'].append(dict(restarted=container_id, timeout=self.parameters.stop_timeout))
+        self.results['actions'].append(dict(restarted=container_id, timeout=self.module.params['stop_timeout']))
         self.results['changed'] = True
         if not self.check_mode:
             try:
-                if self.parameters.stop_timeout:
-                    dummy = self.client.restart(container_id, timeout=self.parameters.stop_timeout)
-                else:
-                    dummy = self.client.restart(container_id)
+                timeout = self.module.params['stop_timeout'] or 10
+                client_timeout = self.client.timeout
+                if client_timeout is not None:
+                    client_timeout += timeout
+                self.client.post_call('/containers/{0}/restart', container_id, params={'t': timeout}, timeout=client_timeout)
             except Exception as exc:
                 self.fail("Error restarting container %s: %s" % (container_id, to_native(exc)))
         return self._get_container(container_id)
 
     def container_stop(self, container_id):
-        if self.parameters.force_kill:
+        if self.param_force_kill:
             self.container_kill(container_id)
             return
-        self.results['actions'].append(dict(stopped=container_id, timeout=self.parameters.stop_timeout))
+        self.results['actions'].append(dict(stopped=container_id, timeout=self.module.params['stop_timeout']))
         self.results['changed'] = True
-        response = None
         if not self.check_mode:
             count = 0
             while True:
                 try:
-                    if self.parameters.stop_timeout:
-                        response = self.client.stop(container_id, timeout=self.parameters.stop_timeout)
+                    timeout = self.module.params['stop_timeout']
+                    if timeout:
+                        params = {'t': timeout}
                     else:
-                        response = self.client.stop(container_id)
+                        params = {}
+                        timeout = 10
+                    client_timeout = self.client.timeout
+                    if client_timeout is not None:
+                        client_timeout += timeout
+                    self.client.post_call('/containers/{0}/stop', container_id, params=params, timeout=client_timeout)
                 except APIError as exc:
                     if 'Unpause the container before stopping or killing' in exc.explanation:
                         # New docker daemon versions do not allow containers to be removed
@@ -3250,7 +2073,7 @@ class ContainerManager(DockerBaseClass):
                         count += 1
                         # Unpause
                         try:
-                            self.client.unpause(container=container_id)
+                            self.client.post_call('/containers/{0}/unpause', container_id)
                         except Exception as exc2:
                             self.fail("Error unpausing container %s for removal: %s" % (container_id, to_native(exc2)))
                         # Now try again
@@ -3260,303 +2083,22 @@ class ContainerManager(DockerBaseClass):
                     self.fail("Error stopping container %s: %s" % (container_id, to_native(exc)))
                 # We only loop when explicitly requested by 'continue'
                 break
-        return response
-
-
-def detect_ipvX_address_usage(client):
-    '''
-    Helper function to detect whether any specified network uses ipv4_address or ipv6_address
-    '''
-    for network in client.module.params.get("networks") or []:
-        if network.get('ipv4_address') is not None or network.get('ipv6_address') is not None:
-            return True
-    return False
-
-
-class AnsibleDockerClientContainer(AnsibleDockerClient):
-    # A list of module options which are not docker container properties
-    __NON_CONTAINER_PROPERTY_OPTIONS = tuple([
-        'env_file', 'force_kill', 'keep_volumes', 'ignore_image', 'name', 'pull', 'purge_networks',
-        'recreate', 'restart', 'state', 'networks', 'cleanup', 'kill_signal',
-        'output_logs', 'paused', 'removal_wait_timeout', 'default_host_ip', 'command_handling',
-    ] + list(DOCKER_COMMON_ARGS.keys()))
-
-    def _parse_comparisons(self):
-        comparisons = {}
-        comp_aliases = {}
-        # Put in defaults
-        explicit_types = dict(
-            command='list',
-            devices='set(dict)',
-            device_requests='set(dict)',
-            dns_search_domains='list',
-            dns_servers='list',
-            env='set',
-            entrypoint='list',
-            etc_hosts='set',
-            mounts='set(dict)',
-            networks='set(dict)',
-            ulimits='set(dict)',
-            device_read_bps='set(dict)',
-            device_write_bps='set(dict)',
-            device_read_iops='set(dict)',
-            device_write_iops='set(dict)',
-        )
-        all_options = set()  # this is for improving user feedback when a wrong option was specified for comparison
-        default_values = dict(
-            stop_timeout='ignore',
-        )
-        for option, data in self.module.argument_spec.items():
-            all_options.add(option)
-            for alias in data.get('aliases', []):
-                all_options.add(alias)
-            # Ignore options which aren't used as container properties
-            if option in self.__NON_CONTAINER_PROPERTY_OPTIONS and option != 'networks':
-                continue
-            # Determine option type
-            if option in explicit_types:
-                datatype = explicit_types[option]
-            elif data['type'] == 'list':
-                datatype = 'set'
-            elif data['type'] == 'dict':
-                datatype = 'dict'
-            else:
-                datatype = 'value'
-            # Determine comparison type
-            if option in default_values:
-                comparison = default_values[option]
-            elif datatype in ('list', 'value'):
-                comparison = 'strict'
-            else:
-                comparison = 'allow_more_present'
-            comparisons[option] = dict(type=datatype, comparison=comparison, name=option)
-            # Keep track of aliases
-            comp_aliases[option] = option
-            for alias in data.get('aliases', []):
-                comp_aliases[alias] = option
-        # Process legacy ignore options
-        if self.module.params['ignore_image']:
-            comparisons['image']['comparison'] = 'ignore'
-        if self.module.params['purge_networks']:
-            comparisons['networks']['comparison'] = 'strict'
-        # Process options
-        if self.module.params.get('comparisons'):
-            # If '*' appears in comparisons, process it first
-            if '*' in self.module.params['comparisons']:
-                value = self.module.params['comparisons']['*']
-                if value not in ('strict', 'ignore'):
-                    self.fail("The wildcard can only be used with comparison modes 'strict' and 'ignore'!")
-                for option, v in comparisons.items():
-                    if option == 'networks':
-                        # `networks` is special: only update if
-                        # some value is actually specified
-                        if self.module.params['networks'] is None:
-                            continue
-                    v['comparison'] = value
-            # Now process all other comparisons.
-            comp_aliases_used = {}
-            for key, value in self.module.params['comparisons'].items():
-                if key == '*':
-                    continue
-                # Find main key
-                key_main = comp_aliases.get(key)
-                if key_main is None:
-                    if key_main in all_options:
-                        self.fail("The module option '%s' cannot be specified in the comparisons dict, "
-                                  "since it does not correspond to container's state!" % key)
-                    self.fail("Unknown module option '%s' in comparisons dict!" % key)
-                if key_main in comp_aliases_used:
-                    self.fail("Both '%s' and '%s' (aliases of %s) are specified in comparisons dict!" % (key, comp_aliases_used[key_main], key_main))
-                comp_aliases_used[key_main] = key
-                # Check value and update accordingly
-                if value in ('strict', 'ignore'):
-                    comparisons[key_main]['comparison'] = value
-                elif value == 'allow_more_present':
-                    if comparisons[key_main]['type'] == 'value':
-                        self.fail("Option '%s' is a value and not a set/list/dict, so its comparison cannot be %s" % (key, value))
-                    comparisons[key_main]['comparison'] = value
-                else:
-                    self.fail("Unknown comparison mode '%s'!" % value)
-        # Add implicit options
-        comparisons['publish_all_ports'] = dict(type='value', comparison='strict', name='published_ports')
-        comparisons['expected_ports'] = dict(type='dict', comparison=comparisons['published_ports']['comparison'], name='expected_ports')
-        comparisons['disable_healthcheck'] = dict(type='value',
-                                                  comparison='ignore' if comparisons['healthcheck']['comparison'] == 'ignore' else 'strict',
-                                                  name='disable_healthcheck')
-        # Check legacy values
-        if self.module.params['ignore_image'] and comparisons['image']['comparison'] != 'ignore':
-            self.module.warn('The ignore_image option has been overridden by the comparisons option!')
-        if self.module.params['purge_networks'] and comparisons['networks']['comparison'] != 'strict':
-            self.module.warn('The purge_networks option has been overridden by the comparisons option!')
-        self.comparisons = comparisons
-
-    def _get_additional_minimal_versions(self):
-        stop_timeout_needed_for_update = self.module.params.get("stop_timeout") is not None and self.module.params.get('state') != 'absent'
-        stop_timeout_supported = self.docker_py_version >= LooseVersion('2.1')
-        if stop_timeout_needed_for_update and not stop_timeout_supported:
-            # We warn (instead of fail) since in older versions, stop_timeout was not used
-            # to update the container's configuration, but only when stopping a container.
-            self.module.warn("Docker SDK for Python's version is %s. Minimum version required is 2.1 to update "
-                             "the container's stop_timeout configuration. "
-                             "If you use the 'docker-py' module, you have to switch to the 'docker' Python package." % (docker_version,))
-        self.option_minimal_versions['stop_timeout']['supported'] = stop_timeout_supported
-
-    def __init__(self, **kwargs):
-        option_minimal_versions = dict(
-            # internal options
-            log_config=dict(),
-            publish_all_ports=dict(),
-            ports=dict(),
-            volume_binds=dict(),
-            name=dict(),
-            # normal options
-            device_read_bps=dict(docker_py_version='1.9.0'),
-            device_read_iops=dict(docker_py_version='1.9.0'),
-            device_write_bps=dict(docker_py_version='1.9.0'),
-            device_write_iops=dict(docker_py_version='1.9.0'),
-            device_requests=dict(docker_py_version='4.3.0', docker_api_version='1.40'),
-            dns_opts=dict(docker_py_version='1.10.0'),
-            auto_remove=dict(docker_py_version='2.1.0'),
-            healthcheck=dict(docker_py_version='2.0.0'),
-            init=dict(docker_py_version='2.2.0'),
-            runtime=dict(docker_py_version='2.4.0'),
-            sysctls=dict(docker_py_version='1.10.0'),
-            userns_mode=dict(docker_py_version='1.10.0'),
-            uts=dict(docker_py_version='3.5.0'),
-            pids_limit=dict(docker_py_version='1.10.0'),
-            mounts=dict(docker_py_version='2.6.0'),
-            cpus=dict(docker_py_version='2.3.0'),
-            storage_opts=dict(docker_py_version='2.1.0'),
-            # specials
-            ipvX_address_supported=dict(docker_py_version='1.9.0',
-                                        detect_usage=detect_ipvX_address_usage,
-                                        usage_msg='ipv4_address or ipv6_address in networks'),
-            stop_timeout=dict(),  # see _get_additional_minimal_versions()
-        )
-
-        super(AnsibleDockerClientContainer, self).__init__(
-            option_minimal_versions=option_minimal_versions,
-            option_minimal_versions_ignore_params=self.__NON_CONTAINER_PROPERTY_OPTIONS,
-            **kwargs
-        )
-
-        self._get_additional_minimal_versions()
-        self._parse_comparisons()
-
-        if self.module.params['container_default_behavior'] == 'compatibility':
-            old_default_values = dict(
-                auto_remove=False,
-                detach=True,
-                init=False,
-                interactive=False,
-                memory="0",
-                paused=False,
-                privileged=False,
-                read_only=False,
-                tty=False,
-            )
-            for param, value in old_default_values.items():
-                if self.module.params[param] is None:
-                    self.module.params[param] = value
 
 
 def main():
     argument_spec = dict(
-        auto_remove=dict(type='bool'),
-        blkio_weight=dict(type='int'),
-        capabilities=dict(type='list', elements='str'),
-        cap_drop=dict(type='list', elements='str'),
-        cgroup_parent=dict(type='str'),
         cleanup=dict(type='bool', default=False),
-        command=dict(type='raw'),
         comparisons=dict(type='dict'),
         container_default_behavior=dict(type='str', default='no_defaults', choices=['compatibility', 'no_defaults']),
-        command_handling=dict(type='str', choices=['compatibility', 'correct'], default='correct'),
-        cpu_period=dict(type='int'),
-        cpu_quota=dict(type='int'),
-        cpus=dict(type='float'),
-        cpuset_cpus=dict(type='str'),
-        cpuset_mems=dict(type='str'),
-        cpu_shares=dict(type='int'),
+        command_handling=dict(type='str', choices=['compatibility', 'correct']),
         default_host_ip=dict(type='str'),
-        detach=dict(type='bool'),
-        devices=dict(type='list', elements='str'),
-        device_read_bps=dict(type='list', elements='dict', options=dict(
-            path=dict(required=True, type='str'),
-            rate=dict(required=True, type='str'),
-        )),
-        device_write_bps=dict(type='list', elements='dict', options=dict(
-            path=dict(required=True, type='str'),
-            rate=dict(required=True, type='str'),
-        )),
-        device_read_iops=dict(type='list', elements='dict', options=dict(
-            path=dict(required=True, type='str'),
-            rate=dict(required=True, type='int'),
-        )),
-        device_write_iops=dict(type='list', elements='dict', options=dict(
-            path=dict(required=True, type='str'),
-            rate=dict(required=True, type='int'),
-        )),
-        device_requests=dict(type='list', elements='dict', options=dict(
-            capabilities=dict(type='list', elements='list'),
-            count=dict(type='int'),
-            device_ids=dict(type='list', elements='str'),
-            driver=dict(type='str'),
-            options=dict(type='dict'),
-        )),
-        dns_servers=dict(type='list', elements='str'),
-        dns_opts=dict(type='list', elements='str'),
-        dns_search_domains=dict(type='list', elements='str'),
-        domainname=dict(type='str'),
-        entrypoint=dict(type='list', elements='str'),
-        env=dict(type='dict'),
-        env_file=dict(type='path'),
-        etc_hosts=dict(type='dict'),
-        exposed_ports=dict(type='list', elements='str', aliases=['exposed', 'expose']),
         force_kill=dict(type='bool', default=False, aliases=['forcekill']),
-        groups=dict(type='list', elements='str'),
-        healthcheck=dict(type='dict', options=dict(
-            test=dict(type='raw'),
-            interval=dict(type='str'),
-            timeout=dict(type='str'),
-            start_period=dict(type='str'),
-            retries=dict(type='int'),
-        )),
-        hostname=dict(type='str'),
         ignore_image=dict(type='bool', default=False),
         image=dict(type='str'),
         image_label_mismatch=dict(type='str', choices=['ignore', 'fail'], default='ignore'),
-        init=dict(type='bool'),
-        interactive=dict(type='bool'),
-        ipc_mode=dict(type='str'),
         keep_volumes=dict(type='bool', default=True),
-        kernel_memory=dict(type='str'),
         kill_signal=dict(type='str'),
-        labels=dict(type='dict'),
-        links=dict(type='list', elements='str'),
-        log_driver=dict(type='str'),
-        log_options=dict(type='dict', aliases=['log_opt']),
-        mac_address=dict(type='str'),
-        memory=dict(type='str'),
-        memory_reservation=dict(type='str'),
-        memory_swap=dict(type='str'),
-        memory_swappiness=dict(type='int'),
-        mounts=dict(type='list', elements='dict', options=dict(
-            target=dict(type='str', required=True),
-            source=dict(type='str'),
-            type=dict(type='str', choices=['bind', 'volume', 'tmpfs', 'npipe'], default='volume'),
-            read_only=dict(type='bool'),
-            consistency=dict(type='str', choices=['default', 'consistent', 'cached', 'delegated']),
-            propagation=dict(type='str', choices=['private', 'rprivate', 'shared', 'rshared', 'slave', 'rslave']),
-            no_copy=dict(type='bool'),
-            labels=dict(type='dict'),
-            volume_driver=dict(type='str'),
-            volume_options=dict(type='dict'),
-            tmpfs_size=dict(type='str'),
-            tmpfs_mode=dict(type='str'),
-        )),
         name=dict(type='str', required=True),
-        network_mode=dict(type='str'),
         networks=dict(type='list', elements='dict', options=dict(
             name=dict(type='str', required=True),
             ipv4_address=dict(type='str'),
@@ -3565,65 +2107,67 @@ def main():
             links=dict(type='list', elements='str'),
         )),
         networks_cli_compatible=dict(type='bool', default=True),
-        oom_killer=dict(type='bool'),
-        oom_score_adj=dict(type='int'),
         output_logs=dict(type='bool', default=False),
         paused=dict(type='bool'),
-        pid_mode=dict(type='str'),
-        pids_limit=dict(type='int'),
-        privileged=dict(type='bool'),
-        publish_all_ports=dict(type='bool'),
-        published_ports=dict(type='list', elements='str', aliases=['ports']),
         pull=dict(type='bool', default=False),
         purge_networks=dict(type='bool', default=False),
-        read_only=dict(type='bool'),
         recreate=dict(type='bool', default=False),
         removal_wait_timeout=dict(type='float'),
         restart=dict(type='bool', default=False),
-        restart_policy=dict(type='str', choices=['no', 'on-failure', 'always', 'unless-stopped']),
-        restart_retries=dict(type='int'),
-        runtime=dict(type='str'),
-        security_opts=dict(type='list', elements='str'),
-        shm_size=dict(type='str'),
         state=dict(type='str', default='started', choices=['absent', 'present', 'started', 'stopped']),
         stop_signal=dict(type='str'),
-        stop_timeout=dict(type='int'),
-        storage_opts=dict(type='dict'),
-        sysctls=dict(type='dict'),
-        tmpfs=dict(type='list', elements='str'),
-        tty=dict(type='bool'),
-        ulimits=dict(type='list', elements='str'),
-        user=dict(type='str'),
-        userns_mode=dict(type='str'),
-        uts=dict(type='str'),
-        volume_driver=dict(type='str'),
-        volumes=dict(type='list', elements='str'),
-        volumes_from=dict(type='list', elements='str'),
-        working_dir=dict(type='str'),
     )
 
+    mutually_exclusive = []
+    required_together = []
+    required_one_of = []
     required_if = [
         ('state', 'present', ['image'])
     ]
+    required_by = {}
 
-    client = AnsibleDockerClientContainer(
+    option_minimal_versions = {}
+
+    active_options = []
+    for options in OPTIONS:
+        if not options.supports_engine('docker_api'):
+            continue
+
+        mutually_exclusive.extend(options.ansible_mutually_exclusive)
+        required_together.extend(options.ansible_required_together)
+        required_one_of.extend(options.ansible_required_one_of)
+        required_if.extend(options.ansible_required_if)
+        required_by.update(options.ansible_required_by)
+        argument_spec.update(options.argument_spec)
+
+        engine = options.get_engine('docker_api')
+        if engine.min_docker_api is not None:
+            for option in options.options:
+                if not option.not_an_ansible_option:
+                    option_minimal_versions[option.name] = {'docker_api_version': engine.min_docker_api}
+
+        active_options.append(options)
+
+    client = AnsibleDockerClient(
         argument_spec=argument_spec,
+        mutually_exclusive=mutually_exclusive,
+        required_together=required_together,
+        required_one_of=required_one_of,
         required_if=required_if,
+        required_by=required_by,
+        option_minimal_versions=option_minimal_versions,
         supports_check_mode=True,
     )
-    if client.module.params['networks_cli_compatible'] is True and client.module.params['networks'] and client.module.params['network_mode'] is None:
-        # Same behavior as Docker CLI: if networks are specified, use the name of the first network as the value for network_mode
-        # (assuming no explicit value is specified for network_mode)
-        client.module.params['network_mode'] = client.module.params['networks'][0]['name']
 
     try:
-        cm = ContainerManager(client)
+        cm = ContainerManager(client.module, client, active_options)
+        cm.run()
         client.module.exit_json(**sanitize_result(cm.results))
     except DockerException as e:
-        client.fail('An unexpected docker error occurred: {0}'.format(to_native(e)), exception=traceback.format_exc())
+        client.fail('An unexpected Docker error occurred: {0}'.format(to_native(e)), exception=traceback.format_exc())
     except RequestException as e:
         client.fail(
-            'An unexpected requests error occurred when Docker SDK for Python tried to talk to the docker daemon: {0}'.format(to_native(e)),
+            'An unexpected requests error occurred when trying to talk to the Docker daemon: {0}'.format(to_native(e)),
             exception=traceback.format_exc())
